refactor eth_call collection
