+++ src/raw_data/historical/eth_calls/execution.rs
@@ -1,3 +1,5 @@
+//! Execution functions for eth_call collection: all process_* functions and multicall infrastructure.
+
 use std::collections::{HashMap, HashSet};
 use std::fs::File;
 use std::path::{Path, PathBuf};
@@ -8,2103 +10,36 @@ use std::time::Instant;
 use alloy::dyn_abi::{DynSolType, DynSolValue};
 use alloy::primitives::{Address, Bytes};
 use alloy::rpc::types::{BlockId, BlockNumberOrTag, TransactionRequest};
-use arrow::array::{
-    Array, ArrayRef, BinaryArray, FixedSizeBinaryArray, FixedSizeBinaryBuilder, UInt32Array,
-    UInt64Array,
-};
-use arrow::datatypes::{DataType, Field, Schema};
-use arrow::record_batch::RecordBatch;
 use futures::{stream, StreamExt, TryStreamExt};
 use parquet::arrow::ArrowWriter;
 use parquet::file::properties::WriterProperties;
-use thiserror::Error;
 use tokio::sync::mpsc::Sender;
 
+use super::config::{compute_function_selector, generate_param_combinations};
+use super::event_triggers::encode_once_call_params;
+use super::frequency::filter_blocks_for_frequency;
+use super::parquet_io::{
+    extract_addresses_from_once_parquet,
+    extract_existing_results_from_parquet, find_null_entries,
+    merge_once_columns, read_existing_once_parquet, read_once_column_index,
+    read_parquet_column_names, write_once_column_index,
+    write_once_results_to_parquet, write_results_to_parquet,
+};
+use super::types::{
+    BlockInfo, BlockRange, CallConfig, CallResult, EthCallCollectionError, FrequencyState, OnceCallConfig,
+    OnceCallResult, TokenCallConfig,
+};
 use crate::decoding::{
-    DecoderMessage, EthCallResult as DecoderEthCallResult,
-    EventCallResult as DecoderEventCallResult, OnceCallResult as DecoderOnceCallResult,
+    DecoderMessage, EthCallResult as DecoderEthCallResult, OnceCallResult as DecoderOnceCallResult,
 };
 use crate::raw_data::historical::factories::FactoryAddressData;
-use crate::raw_data::historical::receipts::{EventTriggerData, LogData};
-use crate::rpc::{RpcError, UnifiedRpcClient};
-use crate::storage::{upload_parquet_to_s3, DataLoader, S3Manifest, StorageManager};
+use crate::rpc::UnifiedRpcClient;
+use crate::storage::{upload_parquet_to_s3, S3Manifest, StorageManager};
 use crate::types::config::contract::{AddressOrAddresses, Contracts};
 use crate::types::config::eth_call::{
-    encode_call_with_params, EthCallConfig, EvmType, Frequency, ParamConfig, ParamError, ParamValue,
+    encode_call_with_params, EthCallConfig, Frequency,
 };
-use crate::types::config::tokens::{AddressOrPoolId, PoolType, Tokens};
 
-#[derive(Debug, Error)]
-pub enum EthCallCollectionError {
-    #[error("RPC error: {0}")]
-    Rpc(#[from] RpcError),
-
-    #[error("IO error: {0}")]
-    Io(#[from] std::io::Error),
-
-    #[error("Parquet error: {0}")]
-    Parquet(#[from] parquet::errors::ParquetError),
-
-    #[error("Arrow error: {0}")]
-    Arrow(#[from] arrow::error::ArrowError),
-
-    #[error("Parameter error: {0}")]
-    Param(#[from] ParamError),
-
-    #[error("Task join error: {0}")]
-    JoinError(String),
-
-    #[error("Event parameter extraction error: {0}")]
-    EventParamExtraction(String),
-}
-
-#[derive(Debug, Clone)]
-pub(crate) struct BlockRange {
-    pub(crate) start: u64,
-    pub(crate) end: u64,
-}
-
-impl BlockRange {
-    pub(crate) fn file_name(&self) -> String {
-        format!("{}-{}.parquet", self.start, self.end - 1)
-    }
-}
-
-#[derive(Debug, Clone)]
-pub(crate) struct BlockInfo {
-    pub(crate) block_number: u64,
-    pub(crate) timestamp: u64,
-}
-
-#[derive(Debug, Clone)]
-pub struct CallConfig {
-    pub contract_name: String,
-    pub address: Address,
-    pub function_name: String,
-    pub encoded_calldata: Bytes,
-    pub param_values: Vec<Vec<u8>>,
-    pub frequency: Frequency,
-    /// Start block for this contract (calls before this block are skipped)
-    pub start_block: Option<u64>,
-}
-
-#[derive(Debug, Clone)]
-pub struct OnceCallConfig {
-    pub function_name: String,
-    pub function_selector: [u8; 4],
-    /// Pre-encoded calldata if no self-address params
-    pub preencoded_calldata: Option<Bytes>,
-    /// Param configs needed when preencoded_calldata is None
-    pub params: Vec<ParamConfig>,
-    /// Optional target address override (resolved from CallTarget)
-    pub target_addresses: Option<Vec<Address>>,
-    /// Start block for this contract (calls before this block are skipped)
-    pub start_block: Option<u64>,
-}
-
-#[derive(Debug)]
-pub(crate) struct OnceCallResult {
-    pub(crate) block_number: u64,
-    pub(crate) block_timestamp: u64,
-    pub(crate) contract_address: [u8; 20],
-    pub(crate) function_results: HashMap<String, Vec<u8>>,
-}
-
-pub struct FrequencyState {
-    pub last_call_times: HashMap<(String, String), u64>,
-}
-
-/// Configuration for an event-triggered call
-#[derive(Debug, Clone)]
-pub struct EventTriggeredCallConfig {
-    /// Contract name or factory collection name
-    pub contract_name: String,
-    /// Target address for the call (None for factory collections - use event emitter)
-    pub target_address: Option<Address>,
-    /// Function name (without params)
-    pub function_name: String,
-    /// Function selector (first 4 bytes of keccak256 of signature)
-    pub function_selector: [u8; 4],
-    /// Parameter configurations
-    pub params: Vec<ParamConfig>,
-    /// Whether this is for a factory collection
-    pub is_factory: bool,
-    /// Start block for this contract (calls before this block are skipped)
-    pub start_block: Option<u64>,
-}
-
-/// Result from an event-triggered call
-#[derive(Debug)]
-pub(crate) struct EventCallResult {
-    pub(crate) block_number: u64,
-    pub(crate) block_timestamp: u64,
-    pub(crate) log_index: u32,
-    pub(crate) target_address: [u8; 20],
-    pub(crate) value_bytes: Vec<u8>,
-    pub(crate) param_values: Vec<Vec<u8>>,
-}
-
-/// Key for grouping event-triggered calls: (source_name, event_signature_hash)
-pub type EventCallKey = (String, [u8; 32]);
-
-#[derive(Debug)]
-pub(crate) struct CallResult {
-    pub(crate) block_number: u64,
-    pub(crate) block_timestamp: u64,
-    pub(crate) contract_address: [u8; 20],
-    pub(crate) value_bytes: Vec<u8>,
-    pub(crate) param_values: Vec<Vec<u8>>,
-}
-
-/// State computed during catchup and passed to the current/streaming phase
-pub(crate) struct EthCallCatchupState {
-    // Immutable configs
-    pub(crate) base_output_dir: PathBuf,
-    pub(crate) range_size: u64,
-    pub(crate) rpc_batch_size: usize,
-    pub(crate) multicall3_address: Option<Address>,
-    pub(crate) call_configs: Vec<CallConfig>,
-    pub(crate) factory_call_configs: HashMap<String, Vec<EthCallConfig>>,
-    pub(crate) event_call_configs: HashMap<EventCallKey, Vec<EventTriggeredCallConfig>>,
-    pub(crate) token_call_configs: Vec<TokenCallConfig>,
-    pub(crate) once_configs: HashMap<String, Vec<OnceCallConfig>>,
-    pub(crate) factory_once_configs: HashMap<String, Vec<OnceCallConfig>>,
-    // Feature flags
-    pub(crate) has_regular_calls: bool,
-    pub(crate) has_once_calls: bool,
-    pub(crate) has_factory_calls: bool,
-    pub(crate) has_factory_once_calls: bool,
-    pub(crate) has_event_triggered_calls: bool,
-    pub(crate) has_token_calls: bool,
-    // Derived constants
-    pub(crate) max_params: usize,
-    pub(crate) factory_max_params: usize,
-    pub(crate) existing_files: HashSet<String>,
-    // S3 manifest for checking remote files
-    pub(crate) s3_manifest: Option<S3Manifest>,
-    // Mutable state carried from catchup
-    pub(crate) factory_addresses: HashMap<String, HashSet<Address>>,
-    pub(crate) frequency_state: FrequencyState,
-    pub(crate) range_data: HashMap<u64, Vec<BlockInfo>>,
-    pub(crate) range_factory_data: HashMap<u64, FactoryAddressData>,
-    pub(crate) range_regular_done: HashSet<u64>,
-    pub(crate) range_factory_done: HashSet<u64>,
-}
-
-// NOTE: build_factory_call_configs was removed - use get_factory_call_configs from factories.rs instead
-
-/// Build event-triggered call configs from contracts configuration
-/// Returns a map from (source_name, event_signature_hash) -> Vec<EventTriggeredCallConfig>
-pub fn build_event_triggered_call_configs(
-    contracts: &Contracts,
-) -> HashMap<EventCallKey, Vec<EventTriggeredCallConfig>> {
-    let mut configs: HashMap<EventCallKey, Vec<EventTriggeredCallConfig>> = HashMap::new();
-
-    for (contract_name, contract) in contracts {
-        // Get contract addresses
-        let addresses = match &contract.address {
-            AddressOrAddresses::Single(addr) => vec![*addr],
-            AddressOrAddresses::Multiple(addrs) => addrs.clone(),
-        };
-        let start_block = contract.start_block.map(|u| u.to::<u64>());
-
-        // Check contract-level calls for on_events frequency
-        if let Some(calls) = &contract.calls {
-            for call in calls {
-                // Iterate over all event trigger configs (supports single or multiple events)
-                for trigger_config in call.frequency.event_configs() {
-                    let event_hash = compute_event_signature_hash(&trigger_config.event);
-                    let key = (trigger_config.source.clone(), event_hash);
-
-                    let selector = compute_function_selector(&call.function);
-                    let function_name = call
-                        .function
-                        .split('(')
-                        .next()
-                        .unwrap_or(&call.function)
-                        .to_string();
-
-                    // Resolve target override if specified, otherwise use contract addresses
-                    let target_addrs = if let Some(target) = &call.target {
-                        match target.resolve_all(contracts) {
-                            Some(addrs) => addrs,
-                            None => {
-                                tracing::warn!(
-                                    "Could not resolve target for event-triggered call {} on contract {}, skipping",
-                                    call.function, contract_name
-                                );
-                                continue;
-                            }
-                        }
-                    } else {
-                        addresses.clone()
-                    };
-
-                    // For each target address, create a config
-                    for address in &target_addrs {
-                        configs
-                            .entry(key.clone())
-                            .or_default()
-                            .push(EventTriggeredCallConfig {
-                                contract_name: contract_name.clone(),
-                                target_address: Some(*address),
-                                function_name: function_name.clone(),
-                                function_selector: selector,
-                                params: call.params.clone(),
-                                is_factory: false,
-                                start_block,
-                            });
-                    }
-                }
-            }
-        }
-
-        // Check factory calls for on_events frequency
-        if let Some(factories) = &contract.factories {
-            for factory in factories {
-                if let Some(calls) = &factory.calls {
-                    for call in calls {
-                        // Iterate over all event trigger configs (supports single or multiple events)
-                        for trigger_config in call.frequency.event_configs() {
-                            let event_hash = compute_event_signature_hash(&trigger_config.event);
-                            let key = (trigger_config.source.clone(), event_hash);
-
-                            let selector = compute_function_selector(&call.function);
-                            let function_name = call
-                                .function
-                                .split('(')
-                                .next()
-                                .unwrap_or(&call.function)
-                                .to_string();
-
-                            // If target is specified, use resolved target; otherwise use event emitter at runtime
-                            let target_address = call.target.as_ref().and_then(|t| {
-                                let resolved = t.resolve(contracts);
-                                if resolved.is_none() {
-                                    tracing::warn!(
-                                        "Could not resolve target for factory event-triggered call {} on collection {}, will use event emitter",
-                                        call.function, factory.collection
-                                    );
-                                }
-                                resolved
-                            });
-
-                            configs.entry(key.clone()).or_default().push(
-                                EventTriggeredCallConfig {
-                                    contract_name: factory.collection.clone(),
-                                    target_address,
-                                    function_name: function_name.clone(),
-                                    function_selector: selector,
-                                    params: call.params.clone(),
-                                    is_factory: true,
-                                    // Factory events use discovery block, not start_block
-                                    start_block: None,
-                                },
-                            );
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    configs
-}
-
-/// Compute the keccak256 hash of an event signature
-pub(crate) fn compute_event_signature_hash(signature: &str) -> [u8; 32] {
-    use alloy::primitives::keccak256;
-    keccak256(signature.as_bytes()).0
-}
-
-/// Extract a parameter value from event data (topics or data fields)
-pub(crate) fn extract_param_from_event(
-    trigger: &EventTriggerData,
-    from_event: &str,
-    param_type: &EvmType,
-) -> Result<(DynSolValue, Vec<u8>), EthCallCollectionError> {
-    let from_event = from_event.trim();
-
-    // Handle "address" - extract the event emitter address
-    if from_event == "address" {
-        // Validate that the expected type is Address
-        if !matches!(param_type, EvmType::Address) {
-            return Err(EthCallCollectionError::EventParamExtraction(format!(
-                "from_event: \"address\" requires param type to be address, got {:?}",
-                param_type
-            )));
-        }
-        let addr = Address::from(trigger.emitter_address);
-        let encoded = DynSolValue::Address(addr).abi_encode();
-        return Ok((DynSolValue::Address(addr), encoded));
-    }
-
-    if let Some(rest) = from_event.strip_prefix("topics[") {
-        // Extract topic index
-        let idx_str = rest.strip_suffix(']').ok_or_else(|| {
-            EthCallCollectionError::EventParamExtraction(format!(
-                "Invalid from_event format: {}",
-                from_event
-            ))
-        })?;
-        let idx: usize = idx_str.parse().map_err(|_| {
-            EthCallCollectionError::EventParamExtraction(format!(
-                "Invalid topic index: {}",
-                idx_str
-            ))
-        })?;
-
-        if idx >= trigger.topics.len() {
-            return Err(EthCallCollectionError::EventParamExtraction(format!(
-                "Topic index {} out of bounds (event has {} topics)",
-                idx,
-                trigger.topics.len()
-            )));
-        }
-
-        let topic_bytes = &trigger.topics[idx];
-        extract_value_from_32_bytes(topic_bytes, param_type)
-    } else if let Some(rest) = from_event.strip_prefix("data[") {
-        // Extract data word index
-        let idx_str = rest.strip_suffix(']').ok_or_else(|| {
-            EthCallCollectionError::EventParamExtraction(format!(
-                "Invalid from_event format: {}",
-                from_event
-            ))
-        })?;
-        let idx: usize = idx_str.parse().map_err(|_| {
-            EthCallCollectionError::EventParamExtraction(format!("Invalid data index: {}", idx_str))
-        })?;
-
-        // Data is stored as 32-byte words
-        let start = idx * 32;
-        let end = start + 32;
-
-        if end > trigger.data.len() {
-            return Err(EthCallCollectionError::EventParamExtraction(format!(
-                "Data index {} out of bounds (event data is {} bytes)",
-                idx,
-                trigger.data.len()
-            )));
-        }
-
-        let mut word = [0u8; 32];
-        word.copy_from_slice(&trigger.data[start..end]);
-        extract_value_from_32_bytes(&word, param_type)
-    } else {
-        Err(EthCallCollectionError::EventParamExtraction(format!(
-            "Unknown from_event format: {} (expected \"address\", \"topics[N]\", or \"data[N]\")",
-            from_event
-        )))
-    }
-}
-
-/// Extract a typed value from a 32-byte word
-pub(crate) fn extract_value_from_32_bytes(
-    bytes: &[u8; 32],
-    param_type: &EvmType,
-) -> Result<(DynSolValue, Vec<u8>), EthCallCollectionError> {
-    use alloy::primitives::{I256, U256};
-
-    match param_type {
-        EvmType::Address => {
-            // Address is stored in the last 20 bytes of the 32-byte word
-            let mut addr_bytes = [0u8; 20];
-            addr_bytes.copy_from_slice(&bytes[12..32]);
-            let addr = Address::from(addr_bytes);
-            let encoded = DynSolValue::Address(addr).abi_encode();
-            Ok((DynSolValue::Address(addr), encoded))
-        }
-        EvmType::Uint256 => {
-            let val = U256::from_be_bytes(*bytes);
-            let encoded = DynSolValue::Uint(val, 256).abi_encode();
-            Ok((DynSolValue::Uint(val, 256), encoded))
-        }
-        EvmType::Uint128 => {
-            let val = U256::from_be_bytes(*bytes);
-            let encoded = DynSolValue::Uint(val, 128).abi_encode();
-            Ok((DynSolValue::Uint(val, 128), encoded))
-        }
-        EvmType::Uint64 => {
-            let val = U256::from_be_bytes(*bytes);
-            let encoded = DynSolValue::Uint(val, 64).abi_encode();
-            Ok((DynSolValue::Uint(val, 64), encoded))
-        }
-        EvmType::Uint32 => {
-            let val = U256::from_be_bytes(*bytes);
-            let encoded = DynSolValue::Uint(val, 32).abi_encode();
-            Ok((DynSolValue::Uint(val, 32), encoded))
-        }
-        EvmType::Uint24 => {
-            let val = U256::from_be_bytes(*bytes);
-            let encoded = DynSolValue::Uint(val, 24).abi_encode();
-            Ok((DynSolValue::Uint(val, 24), encoded))
-        }
-        EvmType::Uint16 => {
-            let val = U256::from_be_bytes(*bytes);
-            let encoded = DynSolValue::Uint(val, 16).abi_encode();
-            Ok((DynSolValue::Uint(val, 16), encoded))
-        }
-        EvmType::Uint8 => {
-            let val = U256::from_be_bytes(*bytes);
-            let encoded = DynSolValue::Uint(val, 8).abi_encode();
-            Ok((DynSolValue::Uint(val, 8), encoded))
-        }
-        EvmType::Uint80 => {
-            let val = U256::from_be_bytes(*bytes);
-            let encoded = DynSolValue::Uint(val, 80).abi_encode();
-            Ok((DynSolValue::Uint(val, 80), encoded))
-        }
-        EvmType::Int256 => {
-            let val = I256::from_be_bytes(*bytes);
-            let encoded = DynSolValue::Int(val, 256).abi_encode();
-            Ok((DynSolValue::Int(val, 256), encoded))
-        }
-        EvmType::Int128 => {
-            let val = I256::from_be_bytes(*bytes);
-            let encoded = DynSolValue::Int(val, 128).abi_encode();
-            Ok((DynSolValue::Int(val, 128), encoded))
-        }
-        EvmType::Int64 => {
-            let val = I256::from_be_bytes(*bytes);
-            let encoded = DynSolValue::Int(val, 64).abi_encode();
-            Ok((DynSolValue::Int(val, 64), encoded))
-        }
-        EvmType::Int32 => {
-            let val = I256::from_be_bytes(*bytes);
-            let encoded = DynSolValue::Int(val, 32).abi_encode();
-            Ok((DynSolValue::Int(val, 32), encoded))
-        }
-        EvmType::Int24 => {
-            let val = I256::from_be_bytes(*bytes);
-            let encoded = DynSolValue::Int(val, 24).abi_encode();
-            Ok((DynSolValue::Int(val, 24), encoded))
-        }
-        EvmType::Int16 => {
-            let val = I256::from_be_bytes(*bytes);
-            let encoded = DynSolValue::Int(val, 16).abi_encode();
-            Ok((DynSolValue::Int(val, 16), encoded))
-        }
-        EvmType::Int8 => {
-            let val = I256::from_be_bytes(*bytes);
-            let encoded = DynSolValue::Int(val, 8).abi_encode();
-            Ok((DynSolValue::Int(val, 8), encoded))
-        }
-        EvmType::Bool => {
-            let val = bytes[31] != 0;
-            let encoded = DynSolValue::Bool(val).abi_encode();
-            Ok((DynSolValue::Bool(val), encoded))
-        }
-        EvmType::Bytes32 => {
-            let b256 = alloy::primitives::B256::from(*bytes);
-            let encoded = DynSolValue::FixedBytes(b256, 32).abi_encode();
-            Ok((DynSolValue::FixedBytes(b256, 32), encoded))
-        }
-        EvmType::Uint160 => {
-            let val = U256::from_be_bytes(*bytes);
-            let encoded = DynSolValue::Uint(val, 160).abi_encode();
-            Ok((DynSolValue::Uint(val, 160), encoded))
-        }
-        EvmType::Uint96 => {
-            let val = U256::from_be_bytes(*bytes);
-            let encoded = DynSolValue::Uint(val, 96).abi_encode();
-            Ok((DynSolValue::Uint(val, 96), encoded))
-        }
-        EvmType::Bytes | EvmType::String => {
-            Err(EthCallCollectionError::EventParamExtraction(format!(
-                "Dynamic types ({:?}) cannot be extracted from event data directly",
-                param_type
-            )))
-        }
-        // Named types delegate to inner type
-        EvmType::Named { inner, .. } => extract_value_from_32_bytes(bytes, inner),
-        // NamedTuple not supported for event param extraction
-        EvmType::NamedTuple(_) => Err(EthCallCollectionError::EventParamExtraction(
-            "NamedTuple cannot be used as a parameter type".to_string(),
-        )),
-        EvmType::UnnamedTuple(_) => Err(EthCallCollectionError::EventParamExtraction(
-            "UnnamedTuple cannot be used as a parameter type".to_string(),
-        )),
-        EvmType::Array(_) => Err(EthCallCollectionError::EventParamExtraction(
-            "Array cannot be used as a parameter type".to_string(),
-        )),
-    }
-}
-
-/// Build parameters for an event-triggered call
-/// Encode calldata for "once" calls that have self-address params
-pub fn encode_once_call_params(
-    function_selector: [u8; 4],
-    param_configs: &[ParamConfig],
-    self_address: Address,
-) -> Result<Bytes, EthCallCollectionError> {
-    let mut params = Vec::with_capacity(param_configs.len());
-
-    for param in param_configs {
-        match param {
-            ParamConfig::Static { param_type, values } => {
-                if let Some(value) = values.first() {
-                    let dyn_val = param_type.parse_value(value)?;
-                    params.push(dyn_val);
-                } else {
-                    return Err(EthCallCollectionError::EventParamExtraction(
-                        "Static param has no values".to_string(),
-                    ));
-                }
-            }
-            ParamConfig::SelfAddress { param_type, source } => {
-                if source == "self" {
-                    let dyn_val = match param_type {
-                        EvmType::Address => DynSolValue::Address(self_address),
-                        _ => {
-                            return Err(EthCallCollectionError::EventParamExtraction(format!(
-                                "source: \"self\" can only be used with address type, got {:?}",
-                                param_type
-                            )));
-                        }
-                    };
-                    params.push(dyn_val);
-                } else {
-                    return Err(EthCallCollectionError::EventParamExtraction(format!(
-                        "Unknown source: {} (expected \"self\")",
-                        source
-                    )));
-                }
-            }
-            ParamConfig::FromEvent { .. } => {
-                return Err(EthCallCollectionError::EventParamExtraction(
-                    "from_event params are not supported for 'once' frequency calls".to_string(),
-                ));
-            }
-        }
-    }
-
-    Ok(encode_call_with_params(function_selector, &params))
-}
-
-pub fn build_event_call_params(
-    trigger: &EventTriggerData,
-    param_configs: &[ParamConfig],
-) -> Result<(Vec<DynSolValue>, Vec<Vec<u8>>), EthCallCollectionError> {
-    let mut params = Vec::with_capacity(param_configs.len());
-    let mut encoded_params = Vec::with_capacity(param_configs.len());
-
-    for param in param_configs {
-        match param {
-            ParamConfig::Static { param_type, values } => {
-                // Use the first static value (event-triggered calls use first value only)
-                if let Some(value) = values.first() {
-                    let dyn_val = param_type.parse_value(value)?;
-                    let encoded = dyn_val.abi_encode();
-                    params.push(dyn_val);
-                    encoded_params.push(encoded);
-                } else {
-                    return Err(EthCallCollectionError::EventParamExtraction(
-                        "Static param has no values".to_string(),
-                    ));
-                }
-            }
-            ParamConfig::FromEvent {
-                param_type,
-                from_event,
-            } => {
-                let (dyn_val, encoded) = extract_param_from_event(trigger, from_event, param_type)?;
-                params.push(dyn_val);
-                encoded_params.push(encoded);
-            }
-            ParamConfig::SelfAddress { param_type, source } => {
-                if source == "self" {
-                    // Use the event emitter address
-                    let addr = Address::from(trigger.emitter_address);
-                    let dyn_val = match param_type {
-                        EvmType::Address => DynSolValue::Address(addr),
-                        _ => {
-                            return Err(EthCallCollectionError::EventParamExtraction(format!(
-                                "source: \"self\" can only be used with address type, got {:?}",
-                                param_type
-                            )));
-                        }
-                    };
-                    let encoded = dyn_val.abi_encode();
-                    params.push(dyn_val);
-                    encoded_params.push(encoded);
-                } else {
-                    return Err(EthCallCollectionError::EventParamExtraction(format!(
-                        "Unknown source: {} (expected \"self\")",
-                        source
-                    )));
-                }
-            }
-        }
-    }
-
-    Ok((params, encoded_params))
-}
-
-/// Process event triggers and make eth_calls
-pub(crate) async fn process_event_triggers(
-    triggers: Vec<EventTriggerData>,
-    event_call_configs: &HashMap<EventCallKey, Vec<EventTriggeredCallConfig>>,
-    factory_addresses: &HashMap<String, HashSet<Address>>,
-    client: &UnifiedRpcClient,
-    output_dir: &Path,
-    rpc_batch_size: usize,
-    decoder_tx: &Option<Sender<DecoderMessage>>,
-    range_start: u64,
-    range_end: u64,
-    chain_name: &str,
-    storage_manager: Option<&Arc<StorageManager>>,
-) -> Result<(), EthCallCollectionError> {
-    // Collect all configured (contract_name, function_name) pairs so we can write empty files
-    // for pairs with no matching events (prevents catchup from retrying)
-    let configured_pairs: HashSet<(String, String)> = event_call_configs
-        .values()
-        .flatten()
-        .map(|c| (c.contract_name.clone(), c.function_name.clone()))
-        .collect();
-
-    if triggers.is_empty() {
-        // Write empty files for all configured pairs
-        for (contract_name, function_name) in &configured_pairs {
-            write_empty_event_call_file(
-                output_dir,
-                contract_name,
-                function_name,
-                range_start,
-                range_end,
-            )?;
-        }
-        return Ok(());
-    }
-
-    // Group triggers by (source_name, event_signature) and then by (contract_name, function_name)
-    // to batch calls together
-    let mut calls_by_output: HashMap<
-        (String, String),
-        Vec<(
-            EventTriggerData,
-            EventTriggeredCallConfig,
-            Address,
-            Vec<DynSolValue>,
-            Vec<Vec<u8>>,
-        )>,
-    > = HashMap::new();
-
-    for trigger in triggers {
-        let key = (trigger.source_name.clone(), trigger.event_signature);
-
-        if let Some(configs) = event_call_configs.get(&key) {
-            for config in configs {
-                // Skip if trigger block is before contract's start_block
-                if let Some(sb) = config.start_block {
-                    if trigger.block_number < sb {
-                        continue;
-                    }
-                }
-
-                // Determine target address
-                let target_address = if let Some(addr) = config.target_address {
-                    addr
-                } else {
-                    // Factory collection - use event emitter, but ONLY if it's a known factory address
-                    let emitter = Address::from(trigger.emitter_address);
-                    match factory_addresses.get(&config.contract_name) {
-                        Some(known_addresses) => {
-                            if !known_addresses.contains(&emitter) {
-                                // Skip - event emitter is not a known factory address
-                                tracing::trace!(
-                                    "Skipping event trigger: emitter {:?} not in known addresses for {}",
-                                    emitter,
-                                    config.contract_name
-                                );
-                                continue;
-                            }
-                            emitter
-                        }
-                        None => {
-                            // No factory addresses known yet for this collection
-                            // Skip this trigger - it will be caught up later when addresses are loaded
-                            // This prevents making calls to random addresses during the race window
-                            tracing::debug!(
-                                "Skipping event trigger for {}: no factory addresses loaded yet (will be processed during catchup)",
-                                config.contract_name
-                            );
-                            continue;
-                        }
-                    }
-                };
-
-                // Build parameters
-                let (params, encoded_params) =
-                    match build_event_call_params(&trigger, &config.params) {
-                        Ok(p) => p,
-                        Err(e) => {
-                            tracing::warn!(
-                                "Failed to build params for {}.{} from event: {}",
-                                config.contract_name,
-                                config.function_name,
-                                e
-                            );
-                            continue;
-                        }
-                    };
-
-                let output_key = (config.contract_name.clone(), config.function_name.clone());
-                calls_by_output.entry(output_key).or_default().push((
-                    trigger.clone(),
-                    config.clone(),
-                    target_address,
-                    params,
-                    encoded_params,
-                ));
-            }
-        }
-    }
-
-    // Track which (contract_name, function_name) pairs got results
-    let mut written_pairs: HashSet<(String, String)> = HashSet::new();
-
-    // Process each group of calls
-    for ((contract_name, function_name), calls) in calls_by_output {
-        if calls.is_empty() {
-            continue;
-        }
-
-        // Mark this pair as having results
-        written_pairs.insert((contract_name.clone(), function_name.clone()));
-
-        // Get block range for output file naming
-        let min_block = calls
-            .iter()
-            .map(|(t, _, _, _, _)| t.block_number)
-            .min()
-            .unwrap();
-        let max_block = calls
-            .iter()
-            .map(|(t, _, _, _, _)| t.block_number)
-            .max()
-            .unwrap();
-
-        tracing::info!(
-            "Processing {} event-triggered eth_calls for {}.{} (blocks {}-{})",
-            calls.len(),
-            contract_name,
-            function_name,
-            min_block,
-            max_block
-        );
-
-        // Build RPC calls
-        let mut pending_calls: Vec<(
-            TransactionRequest,
-            BlockId,
-            &EventTriggerData,
-            Address,
-            Vec<Vec<u8>>,
-        )> = Vec::new();
-
-        for (trigger, config, target_address, params, encoded_params) in &calls {
-            let calldata = encode_call_with_params(config.function_selector, params);
-            let tx = TransactionRequest::default()
-                .to(*target_address)
-                .input(calldata.into());
-            let block_id = BlockId::Number(BlockNumberOrTag::Number(trigger.block_number));
-            pending_calls.push((
-                tx,
-                block_id,
-                trigger,
-                *target_address,
-                encoded_params.clone(),
-            ));
-        }
-
-        // Execute calls in batches with concurrent chunk processing
-        let max_params = calls
-            .iter()
-            .map(|(_, _, _, _, p)| p.len())
-            .max()
-            .unwrap_or(0);
-
-        // Number of chunks to process concurrently
-        let chunk_concurrency = 4;
-
-        // Pre-collect chunks into owned data to avoid lifetime issues
-        let owned_chunks: Vec<
-            Vec<(
-                TransactionRequest,
-                BlockId,
-                EventTriggerData,
-                Address,
-                Vec<Vec<u8>>,
-            )>,
-        > = pending_calls
-            .chunks(rpc_batch_size)
-            .map(|chunk| {
-                chunk
-                    .iter()
-                    .map(|(tx, block_id, trigger, addr, params)| {
-                        (
-                            tx.clone(),
-                            *block_id,
-                            (*trigger).clone(),
-                            *addr,
-                            params.clone(),
-                        )
-                    })
-                    .collect()
-            })
-            .collect();
-
-        let chunk_results: Vec<Vec<EventCallResult>> = stream::iter(owned_chunks)
-            .map(|chunk| {
-                let contract_name = contract_name.clone();
-                let function_name = function_name.clone();
-                async move {
-                    let batch_calls: Vec<(TransactionRequest, BlockId)> = chunk
-                        .iter()
-                        .map(|(tx, block_id, _, _, _)| (tx.clone(), *block_id))
-                        .collect();
-
-                    let results = client.call_batch(batch_calls).await?;
-
-                    let mut chunk_results = Vec::with_capacity(results.len());
-                    for (i, result) in results.into_iter().enumerate() {
-                        let (_, _, trigger, target_address, encoded_params) = &chunk[i];
-
-                        match result {
-                            Ok(bytes) => {
-                                chunk_results.push(EventCallResult {
-                                    block_number: trigger.block_number,
-                                    block_timestamp: trigger.block_timestamp,
-                                    log_index: trigger.log_index,
-                                    target_address: target_address.0.0,
-                                    value_bytes: bytes.to_vec(),
-                                    param_values: encoded_params.clone(),
-                                });
-                            }
-                            Err(e) => {
-                                let params_hex: Vec<String> = encoded_params.iter().map(|p| format!("0x{}", hex::encode(p))).collect();
-                                tracing::warn!(
-                                    "Event-triggered eth_call failed for {}.{} at block {} (address {}, params {:?}): {}",
-                                    contract_name,
-                                    function_name,
-                                    trigger.block_number,
-                                    target_address,
-                                    params_hex,
-                                    e
-                                );
-                                // Skip reverted calls - don't store empty results
-                            }
-                        }
-                    }
-                    Ok::<_, EthCallCollectionError>(chunk_results)
-                }
-            })
-            .buffer_unordered(chunk_concurrency)
-            .try_collect()
-            .await?;
-
-        // Flatten chunk results
-        let mut all_results: Vec<EventCallResult> = chunk_results.into_iter().flatten().collect();
-
-        // Sort by block number, log index
-        all_results.sort_by_key(|r| (r.block_number, r.log_index, r.target_address));
-
-        // Write to parquet and send to decoder
-        if !all_results.is_empty() {
-            let sub_dir = output_dir
-                .join(&contract_name)
-                .join(&function_name)
-                .join("on_events");
-            std::fs::create_dir_all(&sub_dir)?;
-
-            let file_name = format!("{}-{}.parquet", min_block, max_block);
-            let output_path = sub_dir.join(&file_name);
-
-            let result_count = all_results.len();
-            write_event_call_results_to_parquet(&all_results, &output_path, max_params)?;
-
-            tracing::info!(
-                "Wrote {} event-triggered eth_call results to {}",
-                result_count,
-                output_path.display()
-            );
-
-            // Upload to S3 if configured
-            if let Some(sm) = storage_manager {
-                let data_type = format!(
-                    "raw/eth_calls/{}/{}/on_events",
-                    contract_name, function_name
-                );
-                upload_parquet_to_s3(
-                    sm,
-                    &output_path,
-                    chain_name,
-                    &data_type,
-                    min_block,
-                    max_block,
-                )
-                .await
-                .map_err(|e| EthCallCollectionError::Io(std::io::Error::other(e.to_string())))?;
-            }
-
-            // Send to decoder for decoding
-            if let Some(tx) = decoder_tx {
-                let decoder_results: Vec<DecoderEventCallResult> = all_results
-                    .iter()
-                    .map(|r| DecoderEventCallResult {
-                        block_number: r.block_number,
-                        block_timestamp: r.block_timestamp,
-                        log_index: r.log_index,
-                        target_address: r.target_address,
-                        value: r.value_bytes.clone(),
-                    })
-                    .collect();
-
-                let _ = tx
-                    .send(DecoderMessage::EventCallsReady {
-                        range_start: min_block,
-                        range_end: max_block,
-                        contract_name: contract_name.clone(),
-                        function_name: function_name.clone(),
-                        results: decoder_results,
-                        live_mode: false,
-                        retry_transform_after_decode: false,
-                    })
-                    .await;
-            }
-        }
-    }
-
-    // Write empty files for configured pairs that didn't have any matching events
-    for (contract_name, function_name) in configured_pairs {
-        if !written_pairs.contains(&(contract_name.clone(), function_name.clone())) {
-            write_empty_event_call_file(
-                output_dir,
-                &contract_name,
-                &function_name,
-                range_start,
-                range_end,
-            )?;
-        }
-    }
-
-    Ok(())
-}
-
-/// Write an empty parquet file for event-triggered calls when no events matched
-fn write_empty_event_call_file(
-    output_dir: &Path,
-    contract_name: &str,
-    function_name: &str,
-    range_start: u64,
-    range_end: u64,
-) -> Result<(), EthCallCollectionError> {
-    let sub_dir = output_dir
-        .join(contract_name)
-        .join(function_name)
-        .join("on_events");
-    std::fs::create_dir_all(&sub_dir)?;
-
-    let file_name = format!("{}-{}.parquet", range_start, range_end);
-    let output_path = sub_dir.join(&file_name);
-
-    // Write empty parquet with 0 params
-    write_event_call_results_to_parquet(&[], &output_path, 0)?;
-
-    tracing::debug!(
-        "Wrote empty event-triggered eth_call file to {} (no matching events)",
-        output_path.display()
-    );
-
-    Ok(())
-}
-
-/// Process event triggers using Multicall3 aggregate3 to batch all calls per block
-pub(crate) async fn process_event_triggers_multicall(
-    triggers: Vec<EventTriggerData>,
-    event_call_configs: &HashMap<EventCallKey, Vec<EventTriggeredCallConfig>>,
-    factory_addresses: &HashMap<String, HashSet<Address>>,
-    client: &UnifiedRpcClient,
-    output_dir: &Path,
-    rpc_batch_size: usize,
-    decoder_tx: &Option<Sender<DecoderMessage>>,
-    multicall3_address: Address,
-    range_start: u64,
-    range_end: u64,
-    chain_name: &str,
-    storage_manager: Option<&Arc<StorageManager>>,
-) -> Result<(), EthCallCollectionError> {
-    // Collect all configured (contract_name, function_name) pairs so we can write empty files
-    // for pairs with no matching events (prevents catchup from retrying)
-    let configured_pairs: HashSet<(String, String)> = event_call_configs
-        .values()
-        .flatten()
-        .map(|c| (c.contract_name.clone(), c.function_name.clone()))
-        .collect();
-
-    if triggers.is_empty() {
-        // Write empty files for all configured pairs
-        for (contract_name, function_name) in &configured_pairs {
-            write_empty_event_call_file(
-                output_dir,
-                contract_name,
-                function_name,
-                range_start,
-                range_end,
-            )?;
-        }
-        return Ok(());
-    }
-
-    // Group triggers by (source_name, event_signature) and prepare call info
-    // Key: (contract_name, function_name)
-    // Value: Vec<(EventTriggerData, EventTriggeredCallConfig, Address, params, encoded_params)>
-    let mut calls_by_output: HashMap<
-        (String, String),
-        Vec<(
-            EventTriggerData,
-            EventTriggeredCallConfig,
-            Address,
-            Vec<DynSolValue>,
-            Vec<Vec<u8>>,
-        )>,
-    > = HashMap::new();
-
-    for trigger in triggers {
-        let key = (trigger.source_name.clone(), trigger.event_signature);
-
-        if let Some(configs) = event_call_configs.get(&key) {
-            for config in configs {
-                // Skip if trigger block is before contract's start_block
-                if let Some(sb) = config.start_block {
-                    if trigger.block_number < sb {
-                        continue;
-                    }
-                }
-
-                // Determine target address
-                let target_address = if let Some(addr) = config.target_address {
-                    addr
-                } else {
-                    // Factory collection - use event emitter, but ONLY if it's a known factory address
-                    let emitter = Address::from(trigger.emitter_address);
-                    match factory_addresses.get(&config.contract_name) {
-                        Some(known_addresses) => {
-                            if !known_addresses.contains(&emitter) {
-                                continue;
-                            }
-                            emitter
-                        }
-                        None => {
-                            continue;
-                        }
-                    }
-                };
-
-                // Build parameters
-                let (params, encoded_params) =
-                    match build_event_call_params(&trigger, &config.params) {
-                        Ok(p) => p,
-                        Err(e) => {
-                            tracing::warn!(
-                                "Failed to build params for {}.{} from event: {}",
-                                config.contract_name,
-                                config.function_name,
-                                e
-                            );
-                            continue;
-                        }
-                    };
-
-                let output_key = (config.contract_name.clone(), config.function_name.clone());
-                calls_by_output.entry(output_key).or_default().push((
-                    trigger.clone(),
-                    config.clone(),
-                    target_address,
-                    params,
-                    encoded_params,
-                ));
-            }
-        }
-    }
-
-    if calls_by_output.is_empty() {
-        return Ok(());
-    }
-
-    // Group all calls by block number for multicall batching
-    struct EventCallInfo {
-        trigger: EventTriggerData,
-        config: EventTriggeredCallConfig,
-        target_address: Address,
-        params: Vec<DynSolValue>,
-        encoded_params: Vec<Vec<u8>>,
-    }
-
-    let mut calls_by_block: HashMap<u64, Vec<EventCallInfo>> = HashMap::new();
-
-    for ((_contract_name, _function_name), calls) in &calls_by_output {
-        for (trigger, config, target_address, params, encoded_params) in calls {
-            calls_by_block
-                .entry(trigger.block_number)
-                .or_default()
-                .push(EventCallInfo {
-                    trigger: trigger.clone(),
-                    config: config.clone(),
-                    target_address: *target_address,
-                    params: params.clone(),
-                    encoded_params: encoded_params.clone(),
-                });
-        }
-    }
-
-    // Build per-block multicalls with block info included in metadata
-    let mut sorted_blocks: Vec<u64> = calls_by_block.keys().copied().collect();
-    sorted_blocks.sort_unstable();
-
-    let mut block_multicalls: Vec<BlockMulticall<(EventCallMeta, u64, u64)>> = Vec::new();
-
-    for block_number in sorted_blocks {
-        if let Some(calls) = calls_by_block.get(&block_number) {
-            let mut slots: Vec<MulticallSlotGeneric<(EventCallMeta, u64, u64)>> = Vec::new();
-
-            for call_info in calls {
-                let calldata =
-                    encode_call_with_params(call_info.config.function_selector, &call_info.params);
-                slots.push(MulticallSlotGeneric {
-                    block_number,
-                    block_timestamp: call_info.trigger.block_timestamp,
-                    target_address: call_info.target_address,
-                    encoded_calldata: calldata,
-                    metadata: (
-                        EventCallMeta {
-                            contract_name: call_info.config.contract_name.clone(),
-                            function_name: call_info.config.function_name.clone(),
-                            target_address: call_info.target_address,
-                            log_index: call_info.trigger.log_index,
-                            param_values: call_info.encoded_params.clone(),
-                        },
-                        block_number,
-                        call_info.trigger.block_timestamp,
-                    ),
-                });
-            }
-
-            if !slots.is_empty() {
-                block_multicalls.push(BlockMulticall {
-                    block_number,
-                    block_id: BlockId::Number(BlockNumberOrTag::Number(block_number)),
-                    slots,
-                });
-            }
-        }
-    }
-
-    if block_multicalls.is_empty() {
-        return Ok(());
-    }
-
-    // Get min/max blocks for logging and file naming
-    let min_block = block_multicalls
-        .iter()
-        .map(|bm| bm.block_number)
-        .min()
-        .unwrap();
-    let max_block = block_multicalls
-        .iter()
-        .map(|bm| bm.block_number)
-        .max()
-        .unwrap();
-
-    tracing::info!(
-        "Executing {} multicalls for event-triggered calls in blocks {}-{}",
-        block_multicalls.len(),
-        min_block,
-        max_block
-    );
-
-    // Execute all multicalls
-    let results =
-        execute_multicalls_generic(client, multicall3_address, block_multicalls, rpc_batch_size)
-            .await?;
-
-    // Distribute results back to groups
-    let mut group_results: HashMap<(String, String), Vec<EventCallResult>> = HashMap::new();
-
-    for ((meta, block_number, block_timestamp), return_data, _success) in results {
-        let key = (meta.contract_name.clone(), meta.function_name.clone());
-        group_results.entry(key).or_default().push(EventCallResult {
-            block_number,
-            block_timestamp,
-            log_index: meta.log_index,
-            target_address: meta.target_address.0 .0,
-            value_bytes: return_data,
-            param_values: meta.param_values,
-        });
-    }
-
-    // Track which (contract_name, function_name) pairs got results
-    let mut written_pairs: HashSet<(String, String)> = HashSet::new();
-
-    // Write parquet and send to decoder for each group
-    for ((contract_name, function_name), mut results) in group_results {
-        if results.is_empty() {
-            continue;
-        }
-
-        // Mark this pair as having results
-        written_pairs.insert((contract_name.clone(), function_name.clone()));
-
-        // Sort by block number, log index
-        results.sort_by_key(|r| (r.block_number, r.log_index, r.target_address));
-
-        let sub_dir = output_dir
-            .join(&contract_name)
-            .join(&function_name)
-            .join("on_events");
-        std::fs::create_dir_all(&sub_dir)?;
-
-        let file_name = format!("{}-{}.parquet", min_block, max_block);
-        let output_path = sub_dir.join(&file_name);
-
-        let result_count = results.len();
-        let max_params = results
-            .iter()
-            .map(|r| r.param_values.len())
-            .max()
-            .unwrap_or(0);
-        write_event_call_results_to_parquet(&results, &output_path, max_params)?;
-
-        tracing::info!(
-            "Wrote {} multicall event-triggered eth_call results to {}",
-            result_count,
-            output_path.display()
-        );
-
-        // Upload to S3 if configured
-        if let Some(sm) = storage_manager {
-            let data_type = format!(
-                "raw/eth_calls/{}/{}/on_events",
-                contract_name, function_name
-            );
-            upload_parquet_to_s3(
-                sm,
-                &output_path,
-                chain_name,
-                &data_type,
-                min_block,
-                max_block,
-            )
-            .await
-            .map_err(|e| EthCallCollectionError::Io(std::io::Error::other(e.to_string())))?;
-        }
-
-        // Send to decoder
-        if let Some(tx) = decoder_tx {
-            let decoder_results: Vec<DecoderEventCallResult> = results
-                .iter()
-                .map(|r| DecoderEventCallResult {
-                    block_number: r.block_number,
-                    block_timestamp: r.block_timestamp,
-                    log_index: r.log_index,
-                    target_address: r.target_address,
-                    value: r.value_bytes.clone(),
-                })
-                .collect();
-
-            let _ = tx
-                .send(DecoderMessage::EventCallsReady {
-                    range_start: min_block,
-                    range_end: max_block,
-                    contract_name: contract_name.clone(),
-                    function_name: function_name.clone(),
-                    results: decoder_results,
-                    live_mode: false,
-                    retry_transform_after_decode: false,
-                })
-                .await;
-        }
-    }
-
-    // Write empty files for configured pairs that didn't have any matching events
-    for (contract_name, function_name) in configured_pairs {
-        if !written_pairs.contains(&(contract_name.clone(), function_name.clone())) {
-            write_empty_event_call_file(
-                output_dir,
-                &contract_name,
-                &function_name,
-                range_start,
-                range_end,
-            )?;
-        }
-    }
-
-    Ok(())
-}
-
-/// Write event-triggered call results to parquet
-pub(crate) fn write_event_call_results_to_parquet(
-    results: &[EventCallResult],
-    output_path: &Path,
-    num_params: usize,
-) -> Result<(), EthCallCollectionError> {
-    let schema = build_event_call_schema(num_params);
-    let mut arrays: Vec<ArrayRef> = Vec::new();
-
-    // block_number
-    let arr: UInt64Array = results.iter().map(|r| Some(r.block_number)).collect();
-    arrays.push(Arc::new(arr));
-
-    // block_timestamp
-    let arr: UInt64Array = results.iter().map(|r| Some(r.block_timestamp)).collect();
-    arrays.push(Arc::new(arr));
-
-    // log_index
-    let arr: UInt32Array = results.iter().map(|r| Some(r.log_index)).collect();
-    arrays.push(Arc::new(arr));
-
-    // target_address - use FixedSizeBinaryBuilder for empty case compatibility
-    let mut builder = FixedSizeBinaryBuilder::new(20);
-    for r in results {
-        builder.append_value(r.target_address.as_slice())?;
-    }
-    arrays.push(Arc::new(builder.finish()));
-
-    // value
-    let arr: BinaryArray = results
-        .iter()
-        .map(|r| Some(r.value_bytes.as_slice()))
-        .collect();
-    arrays.push(Arc::new(arr));
-
-    // param columns
-    for i in 0..num_params {
-        let arr: BinaryArray = results
-            .iter()
-            .map(|r| r.param_values.get(i).map(|v| v.as_slice()))
-            .collect();
-        arrays.push(Arc::new(arr));
-    }
-
-    let batch = RecordBatch::try_new(schema.clone(), arrays)?;
-
-    let file = File::create(output_path)?;
-    let props = WriterProperties::builder()
-        .set_compression(parquet::basic::Compression::SNAPPY)
-        .build();
-
-    let mut writer = ArrowWriter::try_new(file, schema, Some(props))?;
-    writer.write(&batch)?;
-    writer.close()?;
-
-    Ok(())
-}
-
-/// Build schema for event-triggered call results
-pub(crate) fn build_event_call_schema(num_params: usize) -> Arc<Schema> {
-    let mut fields = vec![
-        Field::new("block_number", DataType::UInt64, false),
-        Field::new("block_timestamp", DataType::UInt64, false),
-        Field::new("log_index", DataType::UInt32, false),
-        Field::new("target_address", DataType::FixedSizeBinary(20), false),
-        Field::new("value", DataType::Binary, false),
-    ];
-
-    for i in 0..num_params {
-        fields.push(Field::new(&format!("param_{}", i), DataType::Binary, true));
-    }
-
-    Arc::new(Schema::new(fields))
-}
-
-/// Load historical factory addresses from parquet files for event trigger catchup
-/// This ensures factory addresses are known before processing historical event triggers
-pub(crate) async fn load_historical_factory_addresses(
-    chain_name: &str,
-    event_call_configs: &HashMap<EventCallKey, Vec<EventTriggeredCallConfig>>,
-    s3_manifest: Option<&S3Manifest>,
-    storage_manager: Option<&Arc<StorageManager>>,
-) -> HashMap<String, HashSet<Address>> {
-    let mut result: HashMap<String, HashSet<Address>> = HashMap::new();
-
-    // Get collection names that need factory addresses (is_factory = true)
-    let factory_collections: HashSet<String> = event_call_configs
-        .values()
-        .flatten()
-        .filter(|c| c.is_factory)
-        .map(|c| c.contract_name.clone())
-        .collect();
-
-    if factory_collections.is_empty() {
-        return result;
-    }
-
-    let factories_dir = PathBuf::from(format!("data/{}/historical/factories", chain_name));
-    let data_loader = DataLoader::new(
-        storage_manager.map(|sm| sm.clone()),
-        chain_name,
-        PathBuf::from("data"),
-    );
-
-    // Collect files to process: local files + S3 manifest ranges
-    let mut files_to_process: HashMap<String, Vec<PathBuf>> = HashMap::new();
-
-    // First, scan local factory directories
-    if let Ok(entries) = std::fs::read_dir(&factories_dir) {
-        for entry in entries.flatten() {
-            let path = entry.path();
-            if !path.is_dir() {
-                continue;
-            }
-
-            let collection_name = path
-                .file_name()
-                .and_then(|s| s.to_str())
-                .unwrap_or("")
-                .to_string();
-
-            // Only load if this collection is needed for event triggers
-            if !factory_collections.contains(&collection_name) {
-                continue;
-            }
-
-            // Read parquet files in this collection
-            if let Ok(files) = std::fs::read_dir(&path) {
-                for file_entry in files.flatten() {
-                    let file_path = file_entry.path();
-                    if !file_path
-                        .extension()
-                        .map(|e| e == "parquet")
-                        .unwrap_or(false)
-                    {
-                        continue;
-                    }
-                    files_to_process
-                        .entry(collection_name.clone())
-                        .or_default()
-                        .push(file_path);
-                }
-            }
-        }
-    }
-
-    // Add S3 manifest factory ranges that aren't already local
-    if let Some(manifest) = s3_manifest {
-        for collection_name in &factory_collections {
-            if let Some(ranges) = manifest.factories.get(collection_name) {
-                let collection_dir = factories_dir.join(collection_name);
-                for &(start, end) in ranges {
-                    let file_name = format!("{}-{}.parquet", start, end);
-                    let file_path = collection_dir.join(&file_name);
-
-                    // Only add if not already in the list
-                    let already_present = files_to_process
-                        .get(collection_name)
-                        .map(|files| files.iter().any(|f| f == &file_path))
-                        .unwrap_or(false);
-
-                    if !already_present {
-                        files_to_process
-                            .entry(collection_name.clone())
-                            .or_default()
-                            .push(file_path);
-                    }
-                }
-            }
-        }
-    }
-
-    // Process all files, downloading from S3 if needed
-    for (collection_name, file_paths) in files_to_process {
-        for file_path in file_paths {
-            // Ensure file is local (download from S3 if needed)
-            match data_loader.ensure_local(&file_path).await {
-                Ok(true) => {
-                    if let Ok(addresses) = read_factory_addresses_from_parquet(&file_path) {
-                        result
-                            .entry(collection_name.clone())
-                            .or_default()
-                            .extend(addresses.into_iter().map(Address::from));
-                    }
-                }
-                Ok(false) => {
-                    tracing::debug!(
-                        "Factory file {} not available locally or in S3",
-                        file_path.display()
-                    );
-                }
-                Err(e) => {
-                    tracing::warn!(
-                        "Failed to ensure factory file {} is local: {}",
-                        file_path.display(),
-                        e
-                    );
-                }
-            }
-        }
-    }
-
-    if !result.is_empty() {
-        tracing::info!(
-            "Loaded historical factory addresses for {} collections",
-            result.len()
-        );
-        for (name, addrs) in &result {
-            tracing::debug!("  {} - {} addresses", name, addrs.len());
-        }
-    }
-
-    result
-}
-
-/// Load factory addresses with block numbers and timestamps for once-call catchup.
-/// Returns a map of collection_name -> Vec<(address, block_number, timestamp)>
-pub(crate) async fn load_factory_addresses_for_once_catchup(
-    chain_name: &str,
-    factory_once_configs: &HashMap<String, Vec<OnceCallConfig>>,
-    s3_manifest: Option<&S3Manifest>,
-    storage_manager: Option<&Arc<StorageManager>>,
-) -> HashMap<String, Vec<(Address, u64, u64)>> {
-    let mut result: HashMap<String, Vec<(Address, u64, u64)>> = HashMap::new();
-
-    let collections_needed: HashSet<String> = factory_once_configs.keys().cloned().collect();
-    if collections_needed.is_empty() {
-        return result;
-    }
-
-    let factories_dir = PathBuf::from(format!("data/{}/historical/factories", chain_name));
-    let data_loader = DataLoader::new(
-        storage_manager.map(|sm| sm.clone()),
-        chain_name,
-        PathBuf::from("data"),
-    );
-
-    // Collect files to process: local files + S3 manifest ranges
-    let mut files_to_process: HashMap<String, Vec<PathBuf>> = HashMap::new();
-
-    // First, scan local factory directories
-    if let Ok(entries) = std::fs::read_dir(&factories_dir) {
-        for entry in entries.flatten() {
-            let path = entry.path();
-            if !path.is_dir() {
-                continue;
-            }
-
-            let collection_name = path
-                .file_name()
-                .and_then(|s| s.to_str())
-                .unwrap_or("")
-                .to_string();
-
-            if !collections_needed.contains(&collection_name) {
-                continue;
-            }
-
-            // Read parquet files in this collection
-            if let Ok(files) = std::fs::read_dir(&path) {
-                for file_entry in files.flatten() {
-                    let file_path = file_entry.path();
-                    if !file_path
-                        .extension()
-                        .map(|e| e == "parquet")
-                        .unwrap_or(false)
-                    {
-                        continue;
-                    }
-                    files_to_process
-                        .entry(collection_name.clone())
-                        .or_default()
-                        .push(file_path);
-                }
-            }
-        }
-    }
-
-    // Add S3 manifest factory ranges that aren't already local
-    if let Some(manifest) = s3_manifest {
-        for collection_name in &collections_needed {
-            if let Some(ranges) = manifest.factories.get(collection_name) {
-                let collection_dir = factories_dir.join(collection_name);
-                for &(start, end) in ranges {
-                    let file_name = format!("{}-{}.parquet", start, end);
-                    let file_path = collection_dir.join(&file_name);
-
-                    // Only add if not already in the list
-                    let already_present = files_to_process
-                        .get(collection_name)
-                        .map(|files| files.iter().any(|f| f == &file_path))
-                        .unwrap_or(false);
-
-                    if !already_present {
-                        files_to_process
-                            .entry(collection_name.clone())
-                            .or_default()
-                            .push(file_path);
-                    }
-                }
-            }
-        }
-    }
-
-    // Process all files, downloading from S3 if needed
-    for (collection_name, file_paths) in files_to_process {
-        for file_path in file_paths {
-            // Ensure file is local (download from S3 if needed)
-            match data_loader.ensure_local(&file_path).await {
-                Ok(true) => {
-                    if let Ok(addresses) = read_factory_addresses_with_blocks(&file_path) {
-                        result
-                            .entry(collection_name.clone())
-                            .or_default()
-                            .extend(addresses);
-                    }
-                }
-                Ok(false) => {
-                    tracing::debug!(
-                        "Factory file {} not available locally or in S3",
-                        file_path.display()
-                    );
-                }
-                Err(e) => {
-                    tracing::warn!(
-                        "Failed to ensure factory file {} is local: {}",
-                        file_path.display(),
-                        e
-                    );
-                }
-            }
-        }
-    }
-
-    if !result.is_empty() {
-        tracing::info!(
-            "Loaded factory addresses for once-call catchup: {} collections",
-            result.len()
-        );
-        for (name, addrs) in &result {
-            tracing::info!("  {} - {} addresses", name, addrs.len());
-        }
-    }
-
-    result
-}
-
-/// Read factory addresses with block number and timestamp from a parquet file
-pub(crate) fn read_factory_addresses_with_blocks(
-    path: &Path,
-) -> Result<Vec<(Address, u64, u64)>, std::io::Error> {
-    use parquet::arrow::arrow_reader::ParquetRecordBatchReaderBuilder;
-
-    let file = File::open(path)?;
-    let reader = match ParquetRecordBatchReaderBuilder::try_new(file) {
-        Ok(builder) => match builder.build() {
-            Ok(r) => r,
-            Err(e) => {
-                tracing::warn!(
-                    "Failed to build parquet reader for {}: {}",
-                    path.display(),
-                    e
-                );
-                return Ok(Vec::new());
-            }
-        },
-        Err(e) => {
-            tracing::warn!(
-                "Failed to create parquet reader for {}: {}",
-                path.display(),
-                e
-            );
-            return Ok(Vec::new());
-        }
-    };
-
-    let mut addresses = Vec::new();
-
-    for batch_result in reader {
-        let batch = match batch_result {
-            Ok(b) => b,
-            Err(e) => {
-                tracing::warn!("Failed to read batch from {}: {}", path.display(), e);
-                continue;
-            }
-        };
-
-        let addr_col_idx = batch.schema().index_of("factory_address").ok();
-        let block_col_idx = batch.schema().index_of("block_number").ok();
-        let ts_col_idx = batch.schema().index_of("block_timestamp").ok();
-
-        if addr_col_idx.is_none() || block_col_idx.is_none() {
-            continue;
-        }
-
-        let addr_col = batch.column(addr_col_idx.unwrap());
-        let block_col = batch.column(block_col_idx.unwrap());
-        let ts_col = ts_col_idx.map(|i| batch.column(i));
-
-        let addr_array = match addr_col.as_any().downcast_ref::<FixedSizeBinaryArray>() {
-            Some(a) => a,
-            None => continue,
-        };
-        let block_array = match block_col.as_any().downcast_ref::<UInt64Array>() {
-            Some(a) => a,
-            None => continue,
-        };
-        let ts_array = ts_col.and_then(|c| c.as_any().downcast_ref::<UInt64Array>());
-
-        for i in 0..addr_array.len() {
-            if addr_array.is_null(i) || block_array.is_null(i) {
-                continue;
-            }
-            let bytes = addr_array.value(i);
-            if bytes.len() != 20 {
-                continue;
-            }
-            let mut addr_bytes = [0u8; 20];
-            addr_bytes.copy_from_slice(bytes);
-            let block_num = block_array.value(i);
-            let timestamp = ts_array.map(|a| a.value(i)).unwrap_or(0);
-            addresses.push((Address::from(addr_bytes), block_num, timestamp));
-        }
-    }
-
-    Ok(addresses)
-}
-
-/// Read factory addresses from a parquet file
-pub(crate) fn read_factory_addresses_from_parquet(
-    path: &Path,
-) -> Result<HashSet<[u8; 20]>, std::io::Error> {
-    use parquet::arrow::arrow_reader::ParquetRecordBatchReaderBuilder;
-
-    let file = File::open(path)?;
-    let reader = match ParquetRecordBatchReaderBuilder::try_new(file) {
-        Ok(builder) => match builder.build() {
-            Ok(r) => r,
-            Err(e) => {
-                tracing::warn!(
-                    "Failed to build parquet reader for {}: {}",
-                    path.display(),
-                    e
-                );
-                return Ok(HashSet::new());
-            }
-        },
-        Err(e) => {
-            tracing::warn!(
-                "Failed to create parquet reader for {}: {}",
-                path.display(),
-                e
-            );
-            return Ok(HashSet::new());
-        }
-    };
-
-    let mut addresses = HashSet::new();
-
-    for batch_result in reader {
-        let batch = match batch_result {
-            Ok(b) => b,
-            Err(e) => {
-                tracing::warn!("Failed to read batch from {}: {}", path.display(), e);
-                continue;
-            }
-        };
-
-        if let Some(col_idx) = batch.schema().index_of("factory_address").ok() {
-            let col = batch.column(col_idx);
-            if let Some(addr_array) = col.as_any().downcast_ref::<FixedSizeBinaryArray>() {
-                for i in 0..addr_array.len() {
-                    if !addr_array.is_null(i) {
-                        let bytes = addr_array.value(i);
-                        if bytes.len() == 20 {
-                            let mut addr = [0u8; 20];
-                            addr.copy_from_slice(bytes);
-                            addresses.insert(addr);
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    Ok(addresses)
-}
-
-/// Existing log range info for catchup
-#[derive(Debug, Clone)]
-pub(crate) struct ExistingLogRange {
-    pub(crate) start: u64,
-    pub(crate) end: u64,
-    pub(crate) file_path: PathBuf,
-}
-
-/// Get existing log file ranges for catchup.
-/// Also includes ranges from S3 manifest that aren't present locally.
-pub(crate) fn get_existing_log_ranges(
-    chain_name: &str,
-    s3_manifest: Option<&S3Manifest>,
-) -> Vec<ExistingLogRange> {
-    let logs_dir = PathBuf::from(format!("data/{}/historical/raw/logs", chain_name));
-    let mut ranges = Vec::new();
-    let mut local_ranges: HashSet<(u64, u64)> = HashSet::new();
-
-    let entries = match std::fs::read_dir(&logs_dir) {
-        Ok(entries) => entries,
-        Err(_) => {
-            // No local directory - just use S3 ranges if available
-            if let Some(manifest) = s3_manifest {
-                for &(start, end) in &manifest.raw_logs {
-                    let file_name = format!("logs_{}-{}.parquet", start, end);
-                    let file_path = logs_dir.join(&file_name);
-                    ranges.push(ExistingLogRange {
-                        start,
-                        end: end + 1, // Convert inclusive end to exclusive
-                        file_path,
-                    });
-                }
-                ranges.sort_by_key(|r| r.start);
-            }
-            return ranges;
-        }
-    };
-
-    for entry in entries.flatten() {
-        let path = entry.path();
-        if !path.extension().map(|e| e == "parquet").unwrap_or(false) {
-            continue;
-        }
-
-        let file_name = match path.file_stem().and_then(|s| s.to_str()) {
-            Some(name) => name,
-            None => continue,
-        };
-
-        // Parse "logs_START-END" format
-        if !file_name.starts_with("logs_") {
-            continue;
-        }
-
-        let range_part = &file_name[5..]; // Skip "logs_"
-        let range_parts: Vec<&str> = range_part.split('-').collect();
-        if range_parts.len() != 2 {
-            continue;
-        }
-
-        let start: u64 = match range_parts[0].parse() {
-            Ok(v) => v,
-            Err(_) => continue,
-        };
-        let end: u64 = match range_parts[1].parse::<u64>() {
-            Ok(v) => v + 1, // Convert inclusive end to exclusive
-            Err(_) => continue,
-        };
-
-        local_ranges.insert((start, end));
-        ranges.push(ExistingLogRange {
-            start,
-            end,
-            file_path: path,
-        });
-    }
-
-    // Add S3-only ranges that aren't present locally
-    if let Some(manifest) = s3_manifest {
-        for &(start, end) in &manifest.raw_logs {
-            let exclusive_end = end + 1; // S3 manifest uses inclusive end
-            if !local_ranges.contains(&(start, exclusive_end)) {
-                let file_name = format!("logs_{}-{}.parquet", start, end);
-                let file_path = logs_dir.join(&file_name);
-                ranges.push(ExistingLogRange {
-                    start,
-                    end: exclusive_end,
-                    file_path,
-                });
-            }
-        }
-    }
-
-    ranges.sort_by_key(|r| r.start);
-    ranges
-}
-
-/// Read logs from a parquet file for catchup
-pub(crate) fn read_logs_from_parquet(
-    file_path: &Path,
-) -> Result<Vec<LogData>, EthCallCollectionError> {
-    use arrow::array::{Array, ListArray};
-    use parquet::arrow::arrow_reader::ParquetRecordBatchReaderBuilder;
-
-    let file = File::open(file_path)?;
-    let reader = ParquetRecordBatchReaderBuilder::try_new(file)?
-        .build()
-        .map_err(|e| EthCallCollectionError::Parquet(e))?;
-
-    let mut logs = Vec::new();
-
-    for batch_result in reader {
-        let batch = batch_result.map_err(EthCallCollectionError::Arrow)?;
-
-        let block_numbers = batch
-            .column_by_name("block_number")
-            .and_then(|c| c.as_any().downcast_ref::<UInt64Array>());
-
-        let block_timestamps = batch
-            .column_by_name("block_timestamp")
-            .and_then(|c| c.as_any().downcast_ref::<UInt64Array>());
-
-        let tx_hashes = batch
-            .column_by_name("transaction_hash")
-            .and_then(|c| c.as_any().downcast_ref::<FixedSizeBinaryArray>());
-
-        let log_indices = batch
-            .column_by_name("log_index")
-            .and_then(|c| c.as_any().downcast_ref::<UInt32Array>());
-
-        let addresses = batch
-            .column_by_name("address")
-            .and_then(|c| c.as_any().downcast_ref::<FixedSizeBinaryArray>());
-
-        let topics_col = batch.column_by_name("topics");
-
-        let data_col = batch
-            .column_by_name("data")
-            .and_then(|c| c.as_any().downcast_ref::<BinaryArray>());
-
-        let (
-            Some(block_numbers),
-            Some(block_timestamps),
-            Some(tx_hashes),
-            Some(log_indices),
-            Some(addresses),
-            Some(data_col),
-        ) = (
-            block_numbers,
-            block_timestamps,
-            tx_hashes,
-            log_indices,
-            addresses,
-            data_col,
-        )
-        else {
-            tracing::warn!(
-                "Missing required columns in log parquet file: {:?}",
-                file_path
-            );
-            continue;
-        };
-
-        for i in 0..batch.num_rows() {
-            let block_number = block_numbers.value(i);
-            let block_timestamp = block_timestamps.value(i);
-
-            let mut tx_hash_bytes = [0u8; 32];
-            tx_hash_bytes.copy_from_slice(tx_hashes.value(i));
-            let transaction_hash = alloy::primitives::B256::from(tx_hash_bytes);
-
-            let log_index = log_indices.value(i);
-
-            let mut address = [0u8; 20];
-            address.copy_from_slice(addresses.value(i));
-
-            // Extract topics from the list column
-            let topics: Vec<[u8; 32]> = if let Some(col) = topics_col {
-                if let Some(list_array) = col.as_any().downcast_ref::<ListArray>() {
-                    let values = list_array.value(i);
-                    if let Some(fixed_array) =
-                        values.as_any().downcast_ref::<FixedSizeBinaryArray>()
-                    {
-                        (0..fixed_array.len())
-                            .map(|j| {
-                                let mut topic = [0u8; 32];
-                                topic.copy_from_slice(fixed_array.value(j));
-                                topic
-                            })
-                            .collect()
-                    } else {
-                        Vec::new()
-                    }
-                } else {
-                    Vec::new()
-                }
-            } else {
-                Vec::new()
-            };
-
-            let data = data_col.value(i).to_vec();
-
-            logs.push(LogData {
-                block_number,
-                block_timestamp,
-                transaction_hash,
-                log_index,
-                address,
-                topics,
-                data,
-            });
-        }
-    }
-
-    Ok(logs)
-}
-
-/// Check if on_events output already exists for a range (locally or in S3)
-pub(crate) fn event_output_exists(
-    output_dir: &Path,
-    contract_name: &str,
-    function_name: &str,
-    range_start: u64,
-    range_end: u64,
-    s3_manifest: Option<&S3Manifest>,
-) -> bool {
-    // Check local first
-    let sub_dir = output_dir
-        .join(contract_name)
-        .join(function_name)
-        .join("on_events");
-    if sub_dir.exists() {
-        // Check for any file that overlaps with this range
-        // Note: range_end is EXCLUSIVE (e.g., 200 means up to block 199)
-        // Output files are named {min_block}-{max_block} where both are INCLUSIVE
-        if let Ok(entries) = std::fs::read_dir(&sub_dir) {
-            for entry in entries.flatten() {
-                let path = entry.path();
-                if !path.extension().map(|e| e == "parquet").unwrap_or(false) {
-                    continue;
-                }
-
-                if let Some(file_stem) = path.file_stem().and_then(|s| s.to_str()) {
-                    let parts: Vec<&str> = file_stem.split('-').collect();
-                    if parts.len() == 2 {
-                        if let (Ok(file_start), Ok(file_end_inclusive)) =
-                            (parts[0].parse::<u64>(), parts[1].parse::<u64>())
-                        {
-                            // Convert to exclusive end for consistent comparison
-                            // Ranges overlap if: start1 < end2 AND start2 < end1 (using exclusive ends)
-                            let file_end_exclusive = file_end_inclusive + 1;
-                            if range_start < file_end_exclusive && file_start < range_end {
-                                return true;
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    // Check S3 manifest
-    // The S3 manifest uses key format "contract/function/on_events" for event-triggered calls
-    // range_end is exclusive but S3 manifest stores inclusive end
-    if let Some(manifest) = s3_manifest {
-        let func_key = format!("{}/on_events", function_name);
-        if manifest.has_raw_eth_calls_granular(contract_name, &func_key, range_start, range_end - 1)
-        {
-            return true;
-        }
-    }
-
-    false
-}
 
 pub(crate) async fn process_factory_range(
     range: &BlockRange,
@@ -2718,439 +653,8 @@ pub(crate) async fn process_factory_range_multicall(
     Ok(())
 }
 
-pub(crate) fn generate_param_combinations(
-    params: &[ParamConfig],
-) -> Result<Vec<Vec<(EvmType, ParamValue, Vec<u8>)>>, ParamError> {
-    if params.is_empty() {
-        return Ok(vec![vec![]]);
-    }
 
-    let mut result = vec![vec![]];
 
-    for param in params {
-        // Only Static params are supported for block-based frequency calls
-        // FromEvent and SelfAddress params are for on_events frequency only
-        let values = match param.values() {
-            Some(values) => values,
-            None => {
-                return Err(ParamError::TypeMismatch {
-                    expected: "static param with values".to_string(),
-                    got: "from_event or source param (only valid for on_events frequency)"
-                        .to_string(),
-                });
-            }
-        };
-
-        let param_type = param.param_type();
-        let mut new_result = Vec::new();
-        for existing in &result {
-            for value in values {
-                let mut combo = existing.clone();
-                let dyn_val = param_type.parse_value(value)?;
-                let encoded = dyn_val.abi_encode();
-                combo.push((param_type.clone(), value.clone(), encoded));
-                new_result.push(combo);
-            }
-        }
-        result = new_result;
-    }
-
-    Ok(result)
-}
-
-pub fn build_call_configs(
-    contracts: &Contracts,
-) -> Result<Vec<CallConfig>, EthCallCollectionError> {
-    let mut configs = Vec::new();
-
-    for (contract_name, contract) in contracts {
-        let default_addresses = match &contract.address {
-            AddressOrAddresses::Single(addr) => vec![*addr],
-            AddressOrAddresses::Multiple(addrs) => addrs.clone(),
-        };
-        let start_block = contract.start_block.map(|u| u.to::<u64>());
-
-        if let Some(calls) = &contract.calls {
-            for call in calls {
-                // Skip once and on_events calls - they're handled separately
-                if call.frequency.is_once() || call.frequency.is_on_events() {
-                    continue;
-                }
-
-                // Resolve target addresses: use target override if specified, otherwise contract addresses
-                let addresses = if let Some(target) = &call.target {
-                    match target.resolve_all(contracts) {
-                        Some(addrs) => addrs,
-                        None => {
-                            tracing::warn!(
-                                "Could not resolve target for call {} on contract {}, skipping",
-                                call.function,
-                                contract_name
-                            );
-                            continue;
-                        }
-                    }
-                } else {
-                    default_addresses.clone()
-                };
-
-                let selector = compute_function_selector(&call.function);
-                let function_name = call
-                    .function
-                    .split('(')
-                    .next()
-                    .unwrap_or(&call.function)
-                    .to_string();
-
-                let param_combinations = generate_param_combinations(&call.params)?;
-
-                for address in &addresses {
-                    for param_combo in &param_combinations {
-                        let dyn_values: Vec<DynSolValue> = param_combo
-                            .iter()
-                            .map(|(param_type, value, _)| param_type.parse_value(value))
-                            .collect::<Result<_, _>>()?;
-
-                        let encoded_calldata = encode_call_with_params(selector, &dyn_values);
-
-                        let param_values: Vec<Vec<u8>> = param_combo
-                            .iter()
-                            .map(|(_, _, encoded)| encoded.clone())
-                            .collect();
-
-                        configs.push(CallConfig {
-                            contract_name: contract_name.clone(),
-                            address: *address,
-                            function_name: function_name.clone(),
-                            encoded_calldata,
-                            param_values,
-                            frequency: call.frequency.clone(),
-                            start_block,
-                        });
-                    }
-                }
-            }
-        }
-    }
-
-    Ok(configs)
-}
-
-/// Configuration for a token pool call
-#[derive(Debug, Clone)]
-#[allow(dead_code)]
-pub(crate) struct TokenCallConfig {
-    /// Token name (used for output directory naming as {token_name}_pool)
-    pub(crate) token_name: String,
-    /// Pool type (v2, v3, v4)
-    pub(crate) pool_type: PoolType,
-    /// Target address for the call (pool address for v2/v3, StateView for v4)
-    pub(crate) target_address: Address,
-    /// Function name (e.g., "slot0")
-    pub(crate) function_name: String,
-    /// Encoded calldata including selector and any params
-    pub(crate) encoded_calldata: Bytes,
-    /// Call frequency
-    pub(crate) frequency: Frequency,
-    /// Output type for decoding
-    pub(crate) output_type: EvmType,
-}
-
-/// Build call configs from token pool configurations
-pub(crate) fn build_token_call_configs(
-    tokens: &Tokens,
-    contracts: &Contracts,
-) -> Result<Vec<TokenCallConfig>, EthCallCollectionError> {
-    let mut configs = Vec::new();
-
-    // Look up UniswapV4StateView address from contracts config
-    let state_view_address = contracts
-        .get("UniswapV4StateView")
-        .and_then(|c| match &c.address {
-            AddressOrAddresses::Single(addr) => Some(*addr),
-            AddressOrAddresses::Multiple(addrs) => addrs.first().copied(),
-        });
-
-    for (token_name, token_config) in tokens {
-        if let Some(pool) = &token_config.pool {
-            if let Some(calls) = &pool.calls {
-                for call in calls {
-                    // Skip once and on_events calls for now
-                    if call.frequency.is_once() || call.frequency.is_on_events() {
-                        continue;
-                    }
-
-                    let selector = compute_function_selector(&call.function);
-                    let function_name = call
-                        .function
-                        .split('(')
-                        .next()
-                        .unwrap_or(&call.function)
-                        .to_string();
-
-                    let (target_address, encoded_calldata) = match pool.pool_type {
-                        PoolType::V2 | PoolType::V3 => {
-                            // Target the pool address directly
-                            let addr = match &pool.address {
-                                AddressOrPoolId::Address(a) => *a,
-                                AddressOrPoolId::PoolId(_) => {
-                                    tracing::warn!(
-                                        "V2/V3 pool {} has PoolId instead of Address, skipping",
-                                        token_name
-                                    );
-                                    continue;
-                                }
-                            };
-                            let calldata = Bytes::copy_from_slice(&selector);
-                            (addr, calldata)
-                        }
-                        PoolType::V4 => {
-                            // Target StateView with pool ID as parameter
-                            let state_view = match state_view_address {
-                                Some(addr) => addr,
-                                None => {
-                                    tracing::warn!(
-                                        "No UniswapV4StateView configured, skipping V4 pool calls for {}",
-                                        token_name
-                                    );
-                                    continue;
-                                }
-                            };
-                            let pool_id_bytes = match &pool.address {
-                                AddressOrPoolId::PoolId(id) => *id,
-                                AddressOrPoolId::Address(_) => {
-                                    tracing::warn!(
-                                        "V4 pool {} has Address instead of PoolId, skipping",
-                                        token_name
-                                    );
-                                    continue;
-                                }
-                            };
-                            // Encode call: selector + abi-encoded pool_id (bytes32)
-                            let pool_id_value = DynSolValue::FixedBytes(pool_id_bytes, 32);
-                            let calldata = encode_call_with_params(selector, &[pool_id_value]);
-                            (state_view, calldata)
-                        }
-                    };
-
-                    configs.push(TokenCallConfig {
-                        token_name: token_name.clone(),
-                        pool_type: pool.pool_type.clone(),
-                        target_address,
-                        function_name,
-                        encoded_calldata,
-                        frequency: call.frequency.clone(),
-                        output_type: call.output_type.clone(),
-                    });
-                }
-            }
-        }
-    }
-
-    Ok(configs)
-}
-
-pub fn build_once_call_configs(contracts: &Contracts) -> HashMap<String, Vec<OnceCallConfig>> {
-    let mut configs: HashMap<String, Vec<OnceCallConfig>> = HashMap::new();
-
-    for (contract_name, contract) in contracts {
-        let start_block = contract.start_block.map(|u| u.to::<u64>());
-
-        if let Some(calls) = &contract.calls {
-            for call in calls {
-                if call.frequency.is_once() {
-                    let selector = compute_function_selector(&call.function);
-                    let function_name = call
-                        .function
-                        .split('(')
-                        .next()
-                        .unwrap_or(&call.function)
-                        .to_string();
-
-                    // Check if call has self-address params (requires dynamic encoding per address)
-                    let (preencoded_calldata, params) = if call.has_self_address_param() {
-                        // Need to encode dynamically per address
-                        (None, call.params.clone())
-                    } else if call.params.is_empty() {
-                        // No params - just the selector
-                        (Some(Bytes::copy_from_slice(&selector)), vec![])
-                    } else {
-                        // Static params only - pre-encode now
-                        // Note: This uses first value from each static param
-                        let mut dyn_values = Vec::new();
-                        let mut all_static = true;
-                        for param in &call.params {
-                            match param {
-                                ParamConfig::Static { param_type, values } => {
-                                    if let Some(value) = values.first() {
-                                        if let Ok(dyn_val) = param_type.parse_value(value) {
-                                            dyn_values.push(dyn_val);
-                                        } else {
-                                            all_static = false;
-                                            break;
-                                        }
-                                    } else {
-                                        all_static = false;
-                                        break;
-                                    }
-                                }
-                                _ => {
-                                    all_static = false;
-                                    break;
-                                }
-                            }
-                        }
-                        if all_static {
-                            (Some(encode_call_with_params(selector, &dyn_values)), vec![])
-                        } else {
-                            // Fallback: store params for dynamic encoding
-                            (None, call.params.clone())
-                        }
-                    };
-
-                    // Resolve target override addresses if specified
-                    let target_addresses = call.target.as_ref().and_then(|t| {
-                        let resolved = t.resolve_all(contracts);
-                        if resolved.is_none() {
-                            tracing::warn!(
-                                "Could not resolve target for once call {} on contract {}, will use contract addresses",
-                                call.function, contract_name
-                            );
-                        }
-                        resolved
-                    });
-
-                    configs
-                        .entry(contract_name.clone())
-                        .or_default()
-                        .push(OnceCallConfig {
-                            function_name,
-                            function_selector: selector,
-                            preencoded_calldata,
-                            params,
-                            target_addresses,
-                            start_block,
-                        });
-                }
-            }
-        }
-    }
-
-    configs
-}
-
-pub fn build_factory_once_call_configs(
-    factory_call_configs: &HashMap<String, Vec<EthCallConfig>>,
-    contracts: &Contracts,
-) -> HashMap<String, Vec<OnceCallConfig>> {
-    let mut configs: HashMap<String, Vec<OnceCallConfig>> = HashMap::new();
-
-    for (collection_name, call_configs) in factory_call_configs {
-        for call in call_configs {
-            if call.frequency.is_once() {
-                let selector = compute_function_selector(&call.function);
-                let function_name = call
-                    .function
-                    .split('(')
-                    .next()
-                    .unwrap_or(&call.function)
-                    .to_string();
-
-                // Resolve target override if specified (resolves to first address only for factory calls)
-                let target_addresses = call.target.as_ref().and_then(|t| {
-                    let resolved = t.resolve(contracts);
-                    if resolved.is_none() {
-                        tracing::warn!(
-                            "Could not resolve target for factory once call {} on collection {}, will use factory addresses",
-                            call.function, collection_name
-                        );
-                    }
-                    // For factory once calls, we resolve to a single target address
-                    resolved.map(|addr| vec![addr])
-                });
-
-                // Check if call has self-address params (requires dynamic encoding per address)
-                let (preencoded_calldata, params) = if call.has_self_address_param() {
-                    // Need to encode dynamically per address
-                    (None, call.params.clone())
-                } else if call.params.is_empty() {
-                    // No params - just the selector
-                    (Some(Bytes::copy_from_slice(&selector)), vec![])
-                } else {
-                    // Static params only - pre-encode now
-                    let mut dyn_values = Vec::new();
-                    let mut all_static = true;
-                    for param in &call.params {
-                        match param {
-                            ParamConfig::Static { param_type, values } => {
-                                if let Some(value) = values.first() {
-                                    if let Ok(dyn_val) = param_type.parse_value(value) {
-                                        dyn_values.push(dyn_val);
-                                    } else {
-                                        all_static = false;
-                                        break;
-                                    }
-                                } else {
-                                    all_static = false;
-                                    break;
-                                }
-                            }
-                            _ => {
-                                all_static = false;
-                                break;
-                            }
-                        }
-                    }
-                    if all_static {
-                        (Some(encode_call_with_params(selector, &dyn_values)), vec![])
-                    } else {
-                        // Fallback: store params for dynamic encoding
-                        (None, call.params.clone())
-                    }
-                };
-
-                configs
-                    .entry(collection_name.clone())
-                    .or_default()
-                    .push(OnceCallConfig {
-                        function_name,
-                        function_selector: selector,
-                        preencoded_calldata,
-                        params,
-                        target_addresses,
-                        // Factory calls use discovery block, not start_block
-                        start_block: None,
-                    });
-            }
-        }
-    }
-
-    configs
-}
-
-pub(crate) fn filter_blocks_for_frequency<'a>(
-    blocks: &'a [BlockInfo],
-    frequency: &Frequency,
-    last_call_timestamp: Option<u64>,
-) -> Vec<&'a BlockInfo> {
-    match frequency {
-        Frequency::EveryBlock => blocks.iter().collect(),
-        Frequency::Once => vec![],        // handled separately
-        Frequency::OnEvents(_) => vec![], // handled separately via event triggers
-        Frequency::EveryNBlocks(n) => blocks.iter().filter(|b| b.block_number % n == 0).collect(),
-        Frequency::Duration(secs) => {
-            let mut result = vec![];
-            let mut last_ts = last_call_timestamp.unwrap_or(0);
-            for block in blocks {
-                if block.timestamp >= last_ts + secs {
-                    result.push(block);
-                    last_ts = block.timestamp;
-                }
-            }
-            result
-        }
-    }
-}
 
 pub(crate) async fn process_once_calls_regular(
     range: &BlockRange,
@@ -4873,14 +2377,6 @@ pub(crate) async fn process_factory_once_calls_multicall(
     Ok(())
 }
 
-pub(crate) fn compute_function_selector(signature: &str) -> [u8; 4] {
-    use alloy::primitives::keccak256;
-    let hash = keccak256(signature.as_bytes());
-    let mut selector = [0u8; 4];
-    selector.copy_from_slice(&hash[0..4]);
-    selector
-}
-
 pub(crate) async fn process_range(
     range: &BlockRange,
     blocks: Vec<BlockInfo>,
@@ -6253,648 +3749,3 @@ pub(crate) async fn process_token_range(
     Ok(())
 }
 
-pub(crate) fn scan_existing_parquet_files(dir: &Path) -> HashSet<String> {
-    let mut files = HashSet::new();
-
-    // Scan nested directories: dir/contract/function/*.parquet
-    if let Ok(contract_entries) = std::fs::read_dir(dir) {
-        for contract_entry in contract_entries.flatten() {
-            let contract_path = contract_entry.path();
-            if !contract_path.is_dir() {
-                continue;
-            }
-            let contract_name = match contract_entry.file_name().to_str() {
-                Some(name) => name.to_string(),
-                None => continue,
-            };
-
-            if let Ok(function_entries) = std::fs::read_dir(&contract_path) {
-                for function_entry in function_entries.flatten() {
-                    let function_path = function_entry.path();
-                    if !function_path.is_dir() {
-                        continue;
-                    }
-                    let function_name = match function_entry.file_name().to_str() {
-                        Some(name) => name.to_string(),
-                        None => continue,
-                    };
-
-                    if let Ok(file_entries) = std::fs::read_dir(&function_path) {
-                        for file_entry in file_entries.flatten() {
-                            if let Some(name) = file_entry.file_name().to_str() {
-                                if name.ends_with(".parquet") {
-                                    // Store as contract/function/filename
-                                    files.insert(format!(
-                                        "{}/{}/{}",
-                                        contract_name, function_name, name
-                                    ));
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-    files
-}
-
-pub(crate) fn build_schema(num_params: usize) -> Arc<Schema> {
-    let mut fields = vec![
-        Field::new("block_number", DataType::UInt64, false),
-        Field::new("block_timestamp", DataType::UInt64, false),
-        Field::new("address", DataType::FixedSizeBinary(20), false),
-        Field::new("value", DataType::Binary, false),
-    ];
-
-    for i in 0..num_params {
-        fields.push(Field::new(&format!("param_{}", i), DataType::Binary, true));
-    }
-
-    Arc::new(Schema::new(fields))
-}
-
-pub(crate) fn write_results_to_parquet(
-    results: &[CallResult],
-    output_path: &Path,
-    num_params: usize,
-) -> Result<(), EthCallCollectionError> {
-    let schema = build_schema(num_params);
-    let mut arrays: Vec<ArrayRef> = Vec::new();
-
-    let arr: UInt64Array = results.iter().map(|r| Some(r.block_number)).collect();
-    arrays.push(Arc::new(arr));
-
-    let arr: UInt64Array = results.iter().map(|r| Some(r.block_timestamp)).collect();
-    arrays.push(Arc::new(arr));
-
-    if results.is_empty() {
-        arrays.push(Arc::new(FixedSizeBinaryBuilder::new(20).finish()));
-    } else {
-        let arr = FixedSizeBinaryArray::try_from_iter(
-            results.iter().map(|r| r.contract_address.as_slice()),
-        )?;
-        arrays.push(Arc::new(arr));
-    }
-
-    let arr: BinaryArray = results
-        .iter()
-        .map(|r| Some(r.value_bytes.as_slice()))
-        .collect();
-    arrays.push(Arc::new(arr));
-
-    for i in 0..num_params {
-        let arr: BinaryArray = results
-            .iter()
-            .map(|r| r.param_values.get(i).map(|v| v.as_slice()))
-            .collect();
-        arrays.push(Arc::new(arr));
-    }
-
-    let batch = RecordBatch::try_new(schema.clone(), arrays)?;
-
-    let file = File::create(output_path)?;
-    let props = WriterProperties::builder()
-        .set_compression(parquet::basic::Compression::SNAPPY)
-        .build();
-
-    let mut writer = ArrowWriter::try_new(file, schema, Some(props))?;
-    writer.write(&batch)?;
-    writer.close()?;
-
-    Ok(())
-}
-
-pub(crate) fn build_once_schema(function_names: &[String]) -> Arc<Schema> {
-    let mut fields = vec![
-        Field::new("block_number", DataType::UInt64, false),
-        Field::new("block_timestamp", DataType::UInt64, false),
-        Field::new("address", DataType::FixedSizeBinary(20), false),
-    ];
-
-    for fn_name in function_names {
-        fields.push(Field::new(
-            &format!("{}_result", fn_name),
-            DataType::Binary,
-            true,
-        ));
-    }
-
-    Arc::new(Schema::new(fields))
-}
-
-/// Read the column index sidecar file from a `once/` directory.
-/// Returns a map of filename -> list of function names whose `{name}_result` columns are present.
-pub(crate) fn read_once_column_index(once_dir: &Path) -> HashMap<String, Vec<String>> {
-    let index_path = once_dir.join("column_index.json");
-    match std::fs::read_to_string(&index_path) {
-        Ok(content) => {
-            let index: HashMap<String, Vec<String>> =
-                serde_json::from_str(&content).unwrap_or_default();
-            tracing::debug!(
-                "Read column index from {}: {} files tracked",
-                index_path.display(),
-                index.len()
-            );
-            index
-        }
-        Err(e) => {
-            tracing::debug!(
-                "No column index at {} ({}), starting fresh",
-                index_path.display(),
-                e.kind()
-            );
-            HashMap::new()
-        }
-    }
-}
-
-/// Write the column index sidecar file to a `once/` directory.
-pub(crate) fn write_once_column_index(
-    once_dir: &Path,
-    index: &HashMap<String, Vec<String>>,
-) -> Result<(), EthCallCollectionError> {
-    let index_path = once_dir.join("column_index.json");
-    tracing::debug!(
-        "Writing column index to {}: {} files tracked",
-        index_path.display(),
-        index.len()
-    );
-    let content = serde_json::to_string_pretty(index).map_err(|e| {
-        std::io::Error::new(
-            std::io::ErrorKind::Other,
-            format!("JSON serialize error: {}", e),
-        )
-    })?;
-    std::fs::write(&index_path, content)?;
-    Ok(())
-}
-
-/// Load or build the column index for a once/ directory.
-/// If the index file exists, load it. Otherwise, scan all parquet files to build it.
-pub(crate) fn load_or_build_once_column_index(once_dir: &Path) -> HashMap<String, Vec<String>> {
-    let index_path = once_dir.join("column_index.json");
-
-    // Try to load existing index
-    if index_path.exists() {
-        if let Ok(content) = std::fs::read_to_string(&index_path) {
-            if let Ok(index) = serde_json::from_str::<HashMap<String, Vec<String>>>(&content) {
-                tracing::debug!(
-                    "Loaded column index from {}: {} files tracked",
-                    index_path.display(),
-                    index.len()
-                );
-                return index;
-            }
-        }
-    }
-
-    // Index doesn't exist or couldn't be loaded - build from parquet files
-    if !once_dir.exists() {
-        return HashMap::new();
-    }
-
-    let mut index = HashMap::new();
-    let parquet_files: Vec<_> = match std::fs::read_dir(once_dir) {
-        Ok(entries) => entries
-            .filter_map(|e| e.ok())
-            .filter(|e| {
-                e.path()
-                    .extension()
-                    .map(|ext| ext == "parquet")
-                    .unwrap_or(false)
-            })
-            .collect(),
-        Err(_) => return HashMap::new(),
-    };
-
-    if parquet_files.is_empty() {
-        return HashMap::new();
-    }
-
-    tracing::info!(
-        "Building column index for {} from {} parquet files",
-        once_dir.display(),
-        parquet_files.len()
-    );
-
-    for entry in parquet_files {
-        let path = entry.path();
-        let file_name = path.file_name().unwrap().to_string_lossy().to_string();
-        let cols: Vec<String> = read_parquet_column_names(&path).into_iter().collect();
-        if !cols.is_empty() {
-            index.insert(file_name, cols);
-        }
-    }
-
-    // Write the newly built index
-    if !index.is_empty() {
-        if let Err(e) = write_once_column_index(once_dir, &index) {
-            tracing::warn!(
-                "Failed to write column index to {}: {}",
-                once_dir.display(),
-                e
-            );
-        } else {
-            tracing::info!(
-                "Built and saved column index for {}: {} files tracked",
-                once_dir.display(),
-                index.len()
-            );
-        }
-    }
-
-    index
-}
-
-/// Read column names from a parquet file's schema (for fallback when index is missing).
-/// Returns function names by stripping the `_result` suffix from column names.
-pub(crate) fn read_parquet_column_names(path: &Path) -> HashSet<String> {
-    use parquet::file::reader::{FileReader, SerializedFileReader};
-
-    let file = match File::open(path) {
-        Ok(f) => f,
-        Err(_) => return HashSet::new(),
-    };
-
-    let reader = match SerializedFileReader::new(file) {
-        Ok(r) => r,
-        Err(_) => return HashSet::new(),
-    };
-
-    let schema = reader.metadata().file_metadata().schema_descr();
-    let mut fn_names = HashSet::new();
-
-    for field in schema.columns() {
-        let name = field.name();
-        // Extract function name from column name (e.g., "getAssetData_result" -> "getAssetData")
-        if let Some(fn_name) = name.strip_suffix("_result") {
-            fn_names.insert(fn_name.to_string());
-        }
-    }
-
-    fn_names
-}
-
-/// Find per-address null entries in `_result` columns.
-/// Returns a map of function_name -> set of addresses that have null values for that function.
-/// Used to identify which specific (address, column) pairs need re-fetching.
-pub(crate) fn find_null_entries(path: &Path) -> HashMap<String, HashSet<[u8; 20]>> {
-    let batches = match read_existing_once_parquet(path) {
-        Ok(b) => b,
-        Err(_) => return HashMap::new(),
-    };
-    if batches.is_empty() {
-        return HashMap::new();
-    }
-    let schema = batches[0].schema();
-    let result_fn_names: Vec<String> = schema
-        .fields()
-        .iter()
-        .filter_map(|f| f.name().strip_suffix("_result").map(|s| s.to_string()))
-        .collect();
-    if result_fn_names.is_empty() {
-        return HashMap::new();
-    }
-
-    let mut null_entries: HashMap<String, HashSet<[u8; 20]>> = HashMap::new();
-    for batch in &batches {
-        let address_col = match batch.column_by_name("address") {
-            Some(col) => col,
-            None => continue,
-        };
-        let address_arr = match address_col.as_any().downcast_ref::<FixedSizeBinaryArray>() {
-            Some(arr) => arr,
-            None => continue,
-        };
-        for fn_name in &result_fn_names {
-            let col_name = format!("{}_result", fn_name);
-            let col = match batch.column_by_name(&col_name) {
-                Some(c) => c,
-                None => continue,
-            };
-            // Skip column entirely if no nulls (O(1) check)
-            if col.null_count() == 0 {
-                continue;
-            }
-            for i in 0..batch.num_rows() {
-                if col.is_null(i) {
-                    let addr_bytes: [u8; 20] = match address_arr.value(i).try_into() {
-                        Ok(b) => b,
-                        Err(_) => continue,
-                    };
-                    null_entries
-                        .entry(fn_name.clone())
-                        .or_default()
-                        .insert(addr_bytes);
-                }
-            }
-        }
-    }
-    null_entries
-}
-
-/// Read an existing once-call parquet file and return all record batches.
-pub(crate) fn read_existing_once_parquet(
-    path: &Path,
-) -> Result<Vec<RecordBatch>, EthCallCollectionError> {
-    use parquet::arrow::arrow_reader::ParquetRecordBatchReaderBuilder;
-    tracing::debug!("Reading existing once parquet from {}", path.display());
-    let file = File::open(path)?;
-    let reader = ParquetRecordBatchReaderBuilder::try_new(file)?.build()?;
-    let batches: Vec<RecordBatch> = reader.collect::<Result<Vec<_>, _>>()?;
-    let total_rows: usize = batches.iter().map(|b| b.num_rows()).sum();
-    let num_cols = batches.first().map(|b| b.num_columns()).unwrap_or(0);
-    tracing::debug!(
-        "Read {} batches ({} rows, {} columns) from {}",
-        batches.len(),
-        total_rows,
-        num_cols,
-        path.display()
-    );
-    Ok(batches)
-}
-
-/// Extract addresses from existing once-call parquet file.
-/// Returns a map of address bytes -> (block_number, timestamp).
-pub(crate) fn extract_addresses_from_once_parquet(
-    batches: &[RecordBatch],
-) -> HashMap<[u8; 20], (u64, u64)> {
-    let mut addresses = HashMap::new();
-
-    for batch in batches {
-        let address_col = match batch.column_by_name("address") {
-            Some(col) => col,
-            None => continue,
-        };
-        let address_arr = match address_col.as_any().downcast_ref::<FixedSizeBinaryArray>() {
-            Some(arr) => arr,
-            None => continue,
-        };
-
-        let block_col = batch.column_by_name("block_number");
-        let timestamp_col = batch.column_by_name("block_timestamp");
-
-        let block_arr = block_col.and_then(|c| c.as_any().downcast_ref::<UInt64Array>());
-        let timestamp_arr = timestamp_col.and_then(|c| c.as_any().downcast_ref::<UInt64Array>());
-
-        for i in 0..batch.num_rows() {
-            let addr_bytes: [u8; 20] = match address_arr.value(i).try_into() {
-                Ok(b) => b,
-                Err(_) => continue,
-            };
-
-            let block_num = block_arr.map(|a| a.value(i)).unwrap_or(0);
-            let timestamp = timestamp_arr.map(|a| a.value(i)).unwrap_or(0);
-
-            addresses
-                .entry(addr_bytes)
-                .or_insert((block_num, timestamp));
-        }
-    }
-
-    addresses
-}
-
-/// Extract existing non-null results for specific function columns from parquet.
-/// Returns a map of address bytes -> function_name -> result bytes.
-/// Only includes addresses that have non-null, non-empty data for at least one of the requested columns.
-pub(crate) fn extract_existing_results_from_parquet(
-    batches: &[RecordBatch],
-    fn_names: &[String],
-) -> HashMap<[u8; 20], HashMap<String, Vec<u8>>> {
-    let mut results: HashMap<[u8; 20], HashMap<String, Vec<u8>>> = HashMap::new();
-
-    for batch in batches {
-        let address_col = match batch.column_by_name("address") {
-            Some(col) => col,
-            None => continue,
-        };
-        let address_arr = match address_col.as_any().downcast_ref::<FixedSizeBinaryArray>() {
-            Some(arr) => arr,
-            None => continue,
-        };
-
-        // Get result columns for each function
-        let result_columns: Vec<(&String, Option<&BinaryArray>)> = fn_names
-            .iter()
-            .map(|fn_name| {
-                let col_name = format!("{}_result", fn_name);
-                let col = batch
-                    .column_by_name(&col_name)
-                    .and_then(|c| c.as_any().downcast_ref::<BinaryArray>());
-                (fn_name, col)
-            })
-            .collect();
-
-        for i in 0..batch.num_rows() {
-            let addr_bytes: [u8; 20] = match address_arr.value(i).try_into() {
-                Ok(b) => b,
-                Err(_) => continue,
-            };
-
-            for (fn_name, maybe_col) in &result_columns {
-                if let Some(col) = maybe_col {
-                    if !col.is_null(i) {
-                        let value = col.value(i);
-                        if !value.is_empty() {
-                            results
-                                .entry(addr_bytes)
-                                .or_default()
-                                .insert((*fn_name).clone(), value.to_vec());
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    results
-}
-
-/// Merge new function result columns into existing once-call record batches.
-/// Matches rows by the `address` column (FixedSizeBinary(20)).
-/// Returns a single merged RecordBatch.
-pub(crate) fn merge_once_columns(
-    existing_batches: &[RecordBatch],
-    new_results: &HashMap<[u8; 20], HashMap<String, Vec<u8>>>,
-    new_fn_names: &[String],
-) -> Result<RecordBatch, EthCallCollectionError> {
-    tracing::debug!(
-        "Merging {} new columns into existing data ({} addresses in new results)",
-        new_fn_names.len(),
-        new_results.len()
-    );
-
-    // Concatenate existing batches into one
-    let existing = arrow::compute::concat_batches(&existing_batches[0].schema(), existing_batches)?;
-
-    let num_rows = existing.num_rows();
-    let address_col = existing
-        .column_by_name("address")
-        .expect("existing once parquet must have address column");
-    let address_arr = address_col
-        .as_any()
-        .downcast_ref::<FixedSizeBinaryArray>()
-        .expect("address column must be FixedSizeBinary(20)");
-
-    tracing::debug!(
-        "Existing data has {} rows, {} columns",
-        num_rows,
-        existing.num_columns()
-    );
-
-    // Build new columns by matching on address
-    let mut new_columns: Vec<ArrayRef> = Vec::new();
-    for fn_name in new_fn_names {
-        let mut matched_count = 0;
-        let arr: BinaryArray = (0..num_rows)
-            .map(|i| {
-                let addr_bytes: [u8; 20] = address_arr.value(i).try_into().unwrap();
-                let result = new_results
-                    .get(&addr_bytes)
-                    .and_then(|m| m.get(fn_name))
-                    .filter(|v| !v.is_empty())
-                    .map(|v| v.as_slice());
-                if result.is_some() {
-                    matched_count += 1;
-                }
-                result
-            })
-            .collect();
-        tracing::debug!(
-            "Column {}_result: matched {}/{} addresses",
-            fn_name,
-            matched_count,
-            num_rows
-        );
-        new_columns.push(Arc::new(arr));
-    }
-
-    // Build merged schema with cell-level merge for existing columns
-    let existing_schema = existing.schema();
-
-    // Map new fn_name -> index in new_columns
-    let new_col_map: HashMap<String, usize> = new_fn_names
-        .iter()
-        .enumerate()
-        .map(|(i, fn_name)| (format!("{}_result", fn_name), i))
-        .collect();
-
-    let mut fields: Vec<Field> = Vec::new();
-    let mut columns: Vec<ArrayRef> = Vec::new();
-    let mut merged_new_indices: HashSet<usize> = HashSet::new();
-
-    // Copy existing columns, doing cell-level merge where new data is available
-    for (i, field) in existing_schema.fields().iter().enumerate() {
-        if let Some(&new_idx) = new_col_map.get(field.name()) {
-            // Column exists in both old and new: cell-level merge
-            // Keep existing non-null values, fill nulls with new data
-            let existing_col = existing.column(i);
-            let new_col = &new_columns[new_idx];
-            let existing_binary = existing_col
-                .as_any()
-                .downcast_ref::<BinaryArray>()
-                .expect("existing result column must be Binary");
-            let new_binary = new_col
-                .as_any()
-                .downcast_ref::<BinaryArray>()
-                .expect("new result column must be Binary");
-            let merged: BinaryArray = (0..existing_binary.len())
-                .map(|row| {
-                    if existing_binary.is_valid(row) {
-                        Some(existing_binary.value(row))
-                    } else if new_binary.is_valid(row) {
-                        Some(new_binary.value(row))
-                    } else {
-                        None
-                    }
-                })
-                .collect();
-            tracing::debug!("Cell-level merged column: {}", field.name());
-            fields.push(field.as_ref().clone());
-            columns.push(Arc::new(merged));
-            merged_new_indices.insert(new_idx);
-        } else {
-            fields.push(field.as_ref().clone());
-            columns.push(existing.column(i).clone());
-        }
-    }
-
-    // Add truly new columns (not present in existing schema)
-    for (i, fn_name) in new_fn_names.iter().enumerate() {
-        if merged_new_indices.contains(&i) {
-            continue;
-        }
-        fields.push(Field::new(
-            &format!("{}_result", fn_name),
-            DataType::Binary,
-            true,
-        ));
-        columns.push(new_columns[i].clone());
-    }
-
-    let merged_schema = Arc::new(Schema::new(fields));
-    let merged = RecordBatch::try_new(merged_schema.clone(), columns)?;
-    tracing::debug!(
-        "Merged result: {} rows, {} columns (schema: {:?})",
-        merged.num_rows(),
-        merged.num_columns(),
-        merged_schema
-            .fields()
-            .iter()
-            .map(|f| f.name())
-            .collect::<Vec<_>>()
-    );
-    Ok(merged)
-}
-
-pub(crate) fn write_once_results_to_parquet(
-    results: &[OnceCallResult],
-    output_path: &Path,
-    function_names: &[String],
-) -> Result<(), EthCallCollectionError> {
-    let schema = build_once_schema(function_names);
-    let mut arrays: Vec<ArrayRef> = Vec::new();
-
-    let arr: UInt64Array = results.iter().map(|r| Some(r.block_number)).collect();
-    arrays.push(Arc::new(arr));
-
-    let arr: UInt64Array = results.iter().map(|r| Some(r.block_timestamp)).collect();
-    arrays.push(Arc::new(arr));
-
-    if results.is_empty() {
-        arrays.push(Arc::new(FixedSizeBinaryBuilder::new(20).finish()));
-    } else {
-        let arr = FixedSizeBinaryArray::try_from_iter(
-            results.iter().map(|r| r.contract_address.as_slice()),
-        )?;
-        arrays.push(Arc::new(arr));
-    }
-
-    for fn_name in function_names {
-        let arr: BinaryArray = results
-            .iter()
-            .map(|r| {
-                r.function_results
-                    .get(fn_name)
-                    .filter(|v| !v.is_empty())
-                    .map(|v| v.as_slice())
-            })
-            .collect();
-        arrays.push(Arc::new(arr));
-    }
-
-    let batch = RecordBatch::try_new(schema.clone(), arrays)?;
-
-    let file = File::create(output_path)?;
-    let props = WriterProperties::builder()
-        .set_compression(parquet::basic::Compression::SNAPPY)
-        .build();
-
-    let mut writer = ArrowWriter::try_new(file, schema, Some(props))?;
-    writer.write(&batch)?;
-    writer.close()?;
-
-    Ok(())
-}
